Refactor frontend components
