+++ modules/server/src/main/js/pages/question.tsx
@@ -1,9 +1,10 @@
 import { observer } from "mobx-react";
-import { Question } from "../components/exercise/question";
 import React, { useEffect, useState } from "react";
 import { container } from "tsyringe";
+import { ExerciseGenerateNextAnswerBtn } from "../components/exercise/exercise-generate-next-answer-btn";
+import { Question } from "../components/exercise/question";
 import { QuestionStore } from "../stores/question-store";
-import { GenerateNextAnswerBtn } from "../components/exercise/generate-next-answer-btn";
+import { QuestionGenerateNextAnswerBtn } from "../components/exercise/question-generate-next-answer-btn";
 
 export const QuestionPage = observer(() => {
     const [question] = useState(() => container.resolve(QuestionStore));
@@ -23,7 +24,7 @@ export const QuestionPage = observer(() => {
         <>
             <Question store={question} showExtendedFeedback />
             <div className="mt-3">
-                <GenerateNextAnswerBtn explicitQuestionStore={question}/>
+                <QuestionGenerateNextAnswerBtn questionStore={question}/>
             </div>
         </>
     );
