+++ modules/server/src/main/js/pages/survey.tsx
@@ -1,17 +1,14 @@
 import { observer } from "mobx-react";
 import React, { useCallback, useEffect, useState } from "react";
-import { GenerateNextQuestionBtn } from "../components/exercise/generate-next-question-btn";
-import { ExerciseHeader } from "../components/exercise/header";
-import { LoadingWrapper } from "../components/common/loader";
-import { Modal } from "../components/common/modal";
-import { GenerateNextAnswerBtn } from "../components/exercise/generate-next-answer-btn";
-import { container } from "tsyringe";
-import { ExerciseStore } from "../stores/exercise-store";
-import { CurrentQuestion } from "../components/exercise/current-question";
-import { Optional } from "../components/common/optional";
-import { useTranslation } from "react-i18next";
 import { Alert } from "react-bootstrap";
+import { useTranslation } from "react-i18next";
+import { container } from "tsyringe";
+import { LoadingWrapper } from "../components/common/loader";
+import { Optional } from "../components/common/optional";
+import { CurrentQuestion } from "../components/exercise/current-question";
+import { ExerciseHeader } from "../components/exercise/header";
 import { SurveyComponent } from "../components/exercise/survey";
+import { ExerciseStore } from "../stores/exercise-store";
 import { Survey } from "../types/survey";
 
 export const SurveyPage = observer(() => {
