+++ modules/server/src/main/resources/static/index.html
@@ -6,5 +6,5 @@
   <link href="/css/vendor.e27e9abceb69f70816bd.css" rel="stylesheet"><link href="/css/client.295f64e0ed406a6427f8.css" rel="stylesheet"></head>
   <body>
     <div id="root"></div>    
-  <script defer src="/js/vendor.0be4e983d10cb40bdb3a.js"></script><script defer src="/js/client.ade84d74256160dd7d7f.js"></script></body>
+  <script defer src="/js/vendor.0be4e983d10cb40bdb3a.js"></script><script defer src="/js/client.9df7ba3840e7cb342f2b.js"></script></body>
 </html>
