--- modules/server/src/main/resources/static/js/client.ade84d74256160dd7d7f.js
@@ -1,1 +1,0 @@
-(()=>{"use strict";var e,t,n,s={558:(e,t,n)=>{var s=n(5338),r=n(6540),i=(n(8630),n(9627)),a=n(2586);function o(e,t){return a.KC([e,a.ch,a.Vx,a.eu("")],null!=t?t:e.name)}function l(e){return new Promise((t=>setTimeout((()=>t()),e)))}function c(e){return null==e}a.KC([a.ch,a.Vx,a.eu("")]);const u=a.NW({attemptId:a.ai,exerciseId:a.ai,questionIds:a.YO(a.ai)},"ExerciseAttempt"),d=o(u),m=a.NW({attemptId:a.ai,questionsCount:a.ai,totalInteractionsCount:a.ai,totalInteractionsWithErrorsCount:a.ai,averageGrade:a.ai},"ExerciseStatisticsItem"),p=a.YO(m);var g=n(8934),h=function(e,t,n,s){return new(n||(n=Promise))((function(r,i){function a(e){try{l(s.next(e))}catch(e){i(e)}}function o(e){try{l(s.throw(e))}catch(e){i(e)}}function l(e){var t;e.done?r(e.value):(t=e.value,t instanceof n?t:new n((function(e){e(t)}))).then(a,o)}l((s=s.apply(e,t||[])).next())}))};const v={method:"GET",headers:{"Content-Type":"application/json"}};function f(e,t,n){return h(this,void 0,void 0,(function*(){const s=Object.assign(Object.assign({},v),{signal:n});return yield E(e,s,t)}))}function b(e,t,n,s){return h(this,void 0,void 0,(function*(){const r=Object.assign(Object.assign({},v),{method:"POST",body:JSON.stringify(t),signal:s});return yield E(e,r,n)}))}function E(e,t,n){return h(this,void 0,void 0,(function*(){var s;const r=yield fetch(e,t).then((e=>h(this,void 0,void 0,(function*(){return e.ok?{status:"ok",payload:n&&n.decode(yield e.json())||a.kX(yield e.json())}:401===e.status?{status:"unauthorized"}:500===e.status?{status:"server_error",payload:yield e.json()}:{status:"unexpected",payload:{message:"Unexpected error"}}})))).catch((e=>({status:"unexpected",payload:{message:"Unexpected error "+e}})));if("ok"===r.status){if(g.isLeft(r.payload)){const e={message:`Type inconsistency for properties of ${null==n?void 0:n.name} type: ${w(r.payload.left).join(", ")}`};return console.error(e),g.left(e)}return g.right(r.payload.right)}return"unauthorized"===r.status?g.left({message:"Unauthorized"}):g.left(null!==(s=r.payload)&&void 0!==s?s:{message:"Unexpected error "})}))}const w=e=>e.map((e=>e.context.map((({key:e})=>e)).join("."))),y=window.location.protocol+"//"+window.location.host,x=a.E$([a.NW({forceNewAttemptCreationEnabled:a.zM,newQuestionGenerationEnabled:a.zM,supplementaryQuestionsEnabled:a.zM,correctAnswerGenerationEnabled:a.zM,preferDecisionTreeBasedSupplementaryEnabled:a.zM}),a.OH({surveyOptions:a.NW({enabled:a.zM,surveyId:a.Yj})})],"ExerciseOptions"),C=a.NW({id:a.ai,options:x},"Exercise");let S=class{getExerciseShortInfo(e){return f(`${y}/api/exercise/shortInfo?id=${e}`,C)}getExerciseAttempt(e){return f(`${y}/api/exercise/getExerciseAttempt?attemptId=${e}`,u)}getExistingExerciseAttempt(e){return f(`${y}/api/exercise/getExistingExerciseAttempt?exerciseId=${e}`,d)}createExerciseAttempt(e){return f(`${y}/api/exercise/createExerciseAttempt?exerciseId=${e}`,u)}createDebugExerciseAttempt(e){return f(`${y}/api/exercise/createDebugExerciseAttempt?exerciseId=${e}`,u)}getExerciseStatistics(e){return f(`${y}/api/exercise/getExerciseStatistics?exerciseId=${e}`,p)}getExercises(){return f(`${y}/api/exercise/getExercises`,a.YO(a.ai))}};S=function(e,t,n,s){var r,i=arguments.length,a=i<3?t:null===s?s=Object.getOwnPropertyDescriptor(t,n):s;if("object"==typeof Reflect&&"function"==typeof Reflect.decorate)a=Reflect.decorate(e,t,n,s);else for(var o=e.length-1;o>=0;o--)(r=e[o])&&(a=(i<3?r(a):i>3?r(t,n,a):r(t,n))||a);return i>3&&a&&Object.defineProperty(t,n,a),a}([(0,i._G)()],S);var _=function(e,t,n,s){return new(n||(n=Promise))((function(r,i){function a(e){try{l(s.next(e))}catch(e){i(e)}}function o(e){try{l(s.throw(e))}catch(e){i(e)}}function l(e){var t;e.done?r(e.value):(t=e.value,t instanceof n?t:new n((function(e){e(t)}))).then(a,o)}l((s=s.apply(e,t||[])).next())}))};let N=class{getCurrentUser(){return _(this,void 0,void 0,(function*(){return g.right({id:999999,displayName:"front user",email:"test@mail.ru",roles:["ADMIN"],language:"EN"})}))}getExerciseShortInfo(e){return _(this,void 0,void 0,(function*(){return g.right({id:-1,options:{forceNewAttemptCreationEnabled:!1,correctAnswerGenerationEnabled:!0,newQuestionGenerationEnabled:!0,supplementaryQuestionsEnabled:!0,preferDecisionTreeBasedSupplementaryEnabled:!1}})}))}getExistingExerciseAttempt(e){return _(this,void 0,void 0,(function*(){return console.log(`getExistingExerciseAttempt?exerciseId=${e}`),g.right("")}))}createExerciseAttempt(e){return _(this,void 0,void 0,(function*(){return console.log(`createExerciseAttempt?exerciseId=${e}`),yield l(3e3),g.right({attemptId:-1,exerciseId:-1,questionIds:[1,2,3,4,5,6,7]})}))}createDebugExerciseAttempt(e){return _(this,void 0,void 0,(function*(){return console.log(`createDebugExerciseAttempt?exerciseId=${e}`),yield l(3e3),g.right({attemptId:-1,exerciseId:-1,questionIds:[1,2,3,4,5,6,7]})}))}getQuestion(e){return _(this,void 0,void 0,(function*(){let t;return console.log(`getQuestion?questionId=${e}`),yield l(3e3),1===e&&(t={type:"SINGLE_CHOICE",questionId:1,text:"question text",answers:[{id:0,text:"answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 "},{id:1,text:"answer2answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 "},{id:2,text:"answer2answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 "}],responses:[],feedback:null,options:{requireContext:!1,showSupplementaryQuestions:!0,displayMode:"radio"}}),2===e&&(t={type:"MULTI_CHOICE",questionId:2,text:"question text",answers:[{id:0,text:"answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 "},{id:1,text:"answer2answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 "},{id:2,text:"answer2answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 answer1 answer1 answer1answer1answer1answer1answer1 answer1answer1 "}],responses:[],feedback:null,options:{requireContext:!1,showSupplementaryQuestions:!0,displayMode:"switch"}}),3===e&&(t={type:"SINGLE_CHOICE",questionId:3,text:'question text with <span id="answer_0">select1</span> and <span id="answer_1">select2</span>',answers:[],responses:[],feedback:null,options:{requireContext:!0,showSupplementaryQuestions:!0,displayMode:"radio"}}),4===e&&(t={type:"MULTI_CHOICE",questionId:4,text:'question text with <span id="answer_0"></span> and <span id="answer_1"></span>',answers:[],responses:[],feedback:null,options:{requireContext:!0,showSupplementaryQuestions:!0,displayMode:"switch"}}),5===e&&(t={type:"MATCHING",questionId:5,text:"question text ",answers:[{id:0,text:"test1"},{id:1,text:"test2"},{id:3,text:"test3"}],groups:[{id:0,text:'<div style="width:70px; height: 40px;">group1<div/>'},{id:1,text:'<div style="width:50px;height: 100px;">group2 group2 group2 group2<div/>'}],responses:[],feedback:null,options:{requireContext:!1,showSupplementaryQuestions:!0,displayMode:"dragNdrop",multipleSelectionEnabled:!0,dropzoneStyle:'{ "display": "inline-block", "minHeight": "40px", "minWidth": "80px" }',dropzoneHtml:"drop",draggableStyle:'{ "padding": "10px", "border": "5px solid", "borderRadius": "5px", "borderColor": "black", "backgroundColor": "white" }'}}),6===e&&(t={type:"MATCHING",questionId:6,text:'question text with <span id="answer_0">drop</span> and <span id="answer_1">drop</span>',answers:[],groups:[{id:0,text:'<div style="width:70px; height: 40px;">group1<div/>'},{id:1,text:'<div style="width:50px;height: 100px;">group2 group2 group2 group2<div/>'}],responses:[],feedback:null,options:{requireContext:!0,showSupplementaryQuestions:!0,displayMode:"dragNdrop",multipleSelectionEnabled:!1,dropzoneStyle:'{ "display": "inline-block", "minHeight": "40px", "minWidth": "80px" }',dropzoneHtml:"drop",draggableStyle:'{ "padding": "10px", "border": "5px solid", "borderRadius": "5px", "borderColor": "black", "backgroundColor": "white" }'}}),7===e&&(t={type:"MULTI_CHOICE",questionId:7,text:'question text with <span id="answer_0"></span> and <span id="answer_1"></span>',answers:[],responses:[],feedback:null,options:{requireContext:!0,showSupplementaryQuestions:!0,displayMode:"dragNdrop",dropzoneStyle:'{ "display": "inline-block", "height": "20px", "width": "20px" }',dropzoneHtml:"",draggableStyle:'{ "height": "20px", "width": "20px" }'}}),t?g.right(t):g.left({message:"No such question"})}))}generateQuestionByAttempt(e){return _(this,void 0,void 0,(function*(){return console.log(`generateQuestionByAttempt?attemptId=${e}`),yield l(3e3),g.left({message:"Method not implemented."})}))}generateQuestionByMetadata(e){return _(this,void 0,void 0,(function*(){return console.log(`generateQuestionByMetadata?metadataId=${e}`),yield l(3e3),g.left({message:"Method not implemented."})}))}generateSupplementaryQuestion(e){return _(this,void 0,void 0,(function*(){return console.log("generateSupplementaryQuestion",e),yield l(3e3),g.left({message:"Method not implemented."})}))}generateNextCorrectAnswer(e){return _(this,void 0,void 0,(function*(){return console.log(`generateNextCorrectAnswer?questionId=${e}`),yield l(3e3),g.left({message:"Method not implemented."})}))}addQuestionAnswer(e){return _(this,void 0,void 0,(function*(){return console.log("addQuestionAnswer",e),yield l(3e3),g.right({isCorrect:!0,grade:.8,correctAnswers:e.answers,correctSteps:1,stepsLeft:1,stepsWithErrors:1,messages:null,strategyDecision:"CONTINUE"})}))}addSupplementaryQuestionAnswer(e){return _(this,void 0,void 0,(function*(){return console.log("addSupplementaryQuestionAnswer",e),yield l(3e3),g.right({message:{type:"SUCCESS",message:"test",violationLaws:[]},action:"CONTINUE_AUTO"})}))}getExerciseStatistics(e){return _(this,void 0,void 0,(function*(){return console.log(`getExerciseStatistics?exerciseId=${e}`),g.left({message:"Method not implemented."})}))}getExercises(){return _(this,void 0,void 0,(function*(){return console.log("getExercises"),g.left({message:"Method not implemented."})}))}getExerciseAttempt(e){return _(this,void 0,void 0,(function*(){return console.log(`getExerciseAttempt?attemptId=${e}`),g.left({message:"Method not implemented."})}))}};N=function(e,t,n,s){var r,i=arguments.length,a=i<3?t:null===s?s=Object.getOwnPropertyDescriptor(t,n):s;if("object"==typeof Reflect&&"function"==typeof Reflect.decorate)a=Reflect.decorate(e,t,n,s);else for(var o=e.length-1;o>=0;o--)(r=e[o])&&(a=(i<3?r(a):i>3?r(t,n,a):r(t,n))||a);return i>3&&a&&Object.defineProperty(t,n,a),a}([(0,i._G)()],N);var A,O=n(7813),I=n(3643),q=function(e,t,n,s){var r,i=arguments.length,a=i<3?t:null===s?s=Object.getOwnPropertyDescriptor(t,n):s;if("object"==typeof Reflect&&"function"==typeof Reflect.decorate)a=Reflect.decorate(e,t,n,s);else for(var o=e.length-1;o>=0;o--)(r=e[o])&&(a=(i<3?r(a):i>3?r(t,n,a):r(t,n))||a);return i>3&&a&&Object.defineProperty(t,n,a),a},k=function(e,t){if("object"==typeof Reflect&&"function"==typeof Reflect.metadata)return Reflect.metadata(e,t)},L=function(e,t,n,s){return new(n||(n=Promise))((function(r,i){function a(e){try{l(s.next(e))}catch(e){i(e)}}function o(e){try{l(s.throw(e))}catch(e){i(e)}}function l(e){var t;e.done?r(e.value):(t=e.value,t instanceof n?t:new n((function(e){e(t)}))).then(a,o)}l((s=s.apply(e,t||[])).next())}))};class j{constructor(e,t){Object.defineProperty(this,"questionController",{enumerable:!0,configurable:!0,writable:!0,value:e}),Object.defineProperty(this,"sourceQuestionId",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"feedback",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"question",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"answer",{enumerable:!0,configurable:!0,writable:!0,value:[]}),Object.defineProperty(this,"questionState",{enumerable:!0,configurable:!0,writable:!0,value:"INITIAL"}),Object.defineProperty(this,"setQuestionState",{enumerable:!0,configurable:!0,writable:!0,value:e=>{this.questionState!==e&&(this.questionState=e)}}),Object.defineProperty(this,"generateSupplementaryQuestion",{enumerable:!0,configurable:!0,writable:!0,value:e=>L(this,void 0,void 0,(function*(){if(0===e.length)throw new Error("violationLaws mist be non-empty");this.setQuestionState("LOADING");const t={questionId:this.sourceQuestionId,violationLaws:e},n=yield this.questionController.generateSupplementaryQuestion(t);(0,O.h5)((()=>{g.isLeft(n)?this.setQuestionState("LOADED"):function(e,t,n,s){if("a"===n&&!s)throw new TypeError("Private accessor was defined without a getter");if("function"==typeof t?e!==t||!s:!t.has(e))throw new TypeError("Cannot read private member from an object whose class did not declare it");return"m"===n?s:"a"===n?s.call(e):s?s.value:t.get(e)}(this,A,"f").call(this,n.right.question,n.right.message)}))}))}),Object.defineProperty(this,"sendAnswers",{enumerable:!0,configurable:!0,writable:!0,value:()=>L(this,void 0,void 0,(function*(){const{question:e}=this;if(!e)throw new Error("Question is empty");const t=(0,O.HO)({questionId:e.questionId,answers:(0,O.HO)([...this.answer])});this.setQuestionState("ANSWER_EVALUATING");const n=yield this.questionController.addSupplementaryQuestionAnswer(t);(0,O.h5)((()=>{g.isLeft(n)?this.setQuestionState("LOADED"):(this.setQuestionState("COMPLETED"),this.feedback=n.right)}))}))}),Object.defineProperty(this,"setAnswer",{enumerable:!0,configurable:!0,writable:!0,value:e=>{this.answer=e}}),A.set(this,((e,t)=>{var n;(null==e?void 0:e.options.requireContext)&&[...e.text.matchAll(/(\<\w.*?\sid\s*?\=([\'\"]))\s*(answer_(\d+?))\2(.*?\>)/gim)].forEach(((t,n)=>{e.text=e.text.replace(t[0],`${t[1]}question_${e.questionId}_${t[3]}_${n}${t[2]} data-answer-id='${t[4]}' ${t[5]}`)})),this.question=null!=e?e:void 0,this.feedback=null!=t?t:void 0,this.answer=null!==(n=null==e?void 0:e.responses)&&void 0!==n?n:[],this.questionState=e?"LOADED":"COMPLETED"})),this.sourceQuestionId=t,(0,O.Gn)(this)}get isQuestionFreezed(){return"LOADED"!==this.questionState}get isFeedbackLoading(){return"ANSWER_EVALUATING"===this.questionState}get canSendQuestionAnswers(){if(!this.question||"COMPLETED"===this.questionState)return!1;switch(this.question.type){case"SINGLE_CHOICE":case"MULTI_CHOICE":return this.answer.length>0;case"ORDER":return!0;case"MATCHING":return this.answer.length===this.question.answers.length;default:return(0,I.absurd)(this.question)}}get questionSubmitMode(){return this.question?"SINGLE_CHOICE"===this.question.type?"IMPLICIT":"EXPLICIT":null}}A=new WeakMap,q([O.sH,k("design:type",Number)],j.prototype,"sourceQuestionId",void 0),q([O.sH,k("design:type",Object)],j.prototype,"feedback",void 0),q([O.sH,k("design:type",Object)],j.prototype,"question",void 0),q([O.sH,k("design:type",Array)],j.prototype,"answer",void 0),q([O.sH,k("design:type",String)],j.prototype,"questionState",void 0),q([O.XI,k("design:type",Object)],j.prototype,"setQuestionState",void 0),q([O.EW,k("design:type",Object),k("design:paramtypes",[])],j.prototype,"isQuestionFreezed",null),q([O.EW,k("design:type",Object),k("design:paramtypes",[])],j.prototype,"isFeedbackLoading",null),q([O.EW,k("design:type",Boolean),k("design:paramtypes",[])],j.prototype,"canSendQuestionAnswers",null),q([O.EW,k("design:type",Object),k("design:paramtypes",[])],j.prototype,"questionSubmitMode",null),q([O.XI,k("design:type",Object)],j.prototype,"generateSupplementaryQuestion",void 0),q([O.XI,k("design:type",Object)],j.prototype,"sendAnswers",void 0),q([O.XI,k("design:type",Object)],j.prototype,"setAnswer",void 0);const Q=a.E$([a.NW({answer:a.PV([a.ai,a.ai]),isСreatedByUser:a.zM}),a.OH({createdByInteraction:a.KC([a.ai,a.ch])})]),T=a.NW({name:a.Yj,canCreateSupplementaryQuestion:a.zM}),P=a.KC([a.NW({type:a.eu("SUCCESS"),message:a.Yj,violationLaws:a.KC([a.YO(T),a.ch])}),a.NW({type:a.eu("ERROR"),message:a.Yj,violationLaws:a.KC([a.YO(T),a.ch])})]),D=a.E$([a.NW({isCorrect:a.zM}),a.OH({isCorrect:a.zM,grade:a.KC([a.ai,a.ch]),violations:a.KC([a.YO(a.ai),a.ch]),correctAnswers:a.KC([a.YO(Q),a.ch]),correctSteps:a.KC([a.ai,a.ch]),stepsLeft:a.KC([a.ai,a.ch]),stepsWithErrors:a.KC([a.ai,a.ch]),message:a.KC([a.YO(P),a.ch]),strategyDecision:a.KC([a.ZE({CONTINUE:null,FINISH:null}),a.ch])})],"Feedback"),R=a.E$([D,a.OH({trace:a.KC([a.YO(a.Yj),a.ch])})]),M=a.NW({requireContext:a.zM,showSupplementaryQuestions:a.zM},"QuestionOptions"),$=a.E$([M,a.NW({showTrace:a.zM,multipleSelectionEnabled:a.zM,requireAllAnswers:a.zM}),a.OH({orderNumberOptions:a.E$([a.NW({delimiter:a.Yj,position:a.ZE({PREFIX:null,SUFFIX:null,BOTTOM:null,NONE:null})}),a.OH({replacers:a.KC([a.YO(a.Yj),a.ch])})])})],"OrderQuestionOptions"),H=a.E$([M,a.NW({multipleSelectionEnabled:a.zM}),a.KC([a.NW({displayMode:a.eu("combobox")}),a.NW({displayMode:a.eu("dragNdrop"),draggableStyle:a.Yj,dropzoneStyle:a.Yj,dropzoneHtml:a.Yj})])],"MatchingQuestionOptions"),F=a.E$([M,a.NW({displayMode:a.ZE({radio:null,dragNdrop:null})})],"SingleChoiceQuestionOptions"),V=a.E$([M,a.KC([a.E$([a.NW({displayMode:a.eu("switch")}),a.OH({selectorReplacers:a.KC([a.PV([a.Yj,a.Yj]),a.ch])})]),a.NW({displayMode:a.eu("dragNdrop"),dropzoneHtml:a.Yj,dropzoneStyle:a.Yj,draggableStyle:a.Yj})])],"MultiChoiceQuestionOptions"),G=a.ZE({SINGLE_CHOICE:null,MULTI_CHOICE:null,MATCHING:null,ORDER:null},"QuestionType"),B=a.Yj,Y=a.NW({id:a.ai,text:B},"QuestionAnswer"),z=a.NW({questionId:a.ai,type:G,options:M,text:B,answers:a.YO(Y),responses:a.KC([a.YO(Q),a.ch]),feedback:a.KC([D,a.ch])},"QuestionBase"),W=a.E$([z,a.NW({type:a.eu("ORDER"),options:$,feedback:a.KC([R,a.ch])}),a.OH({initialTrace:a.KC([a.YO(a.Yj),a.ch])})],"OrderQuestion"),U=a.E$([z,a.NW({type:a.eu("SINGLE_CHOICE"),options:F})],"SingleChoiceQuestion"),X=a.E$([z,a.NW({type:a.eu("MULTI_CHOICE"),options:V})],"MultiChoiceQuestion"),K=a.E$([z,a.NW({type:a.eu("MATCHING"),answers:a.YO(Y),groups:a.YO(Y),options:H})],"MatchingQuestion"),Z=a.KC([W,U,X,K],"Question"),J=(o(Z),a.ZE({CONTINUE_AUTO:null,CONTINUE_MANUAL:null,FINISH:null})),ee=a.NW({message:P,action:J}),te=a.OH({question:a.KC([Z,a.ch]),message:a.KC([ee,a.ch])});let ne=class{generateQuestionByAttempt(e){return f(`${y}/api/question/generate?attemptId=${e}`,Z)}generateQuestionByMetadata(e){return f(`${y}/api/question/generateByMetadata?metadataId=${e}`,Z)}getQuestion(e){return f(`${y}/api/question?questionId=${e}`,Z)}generateSupplementaryQuestion(e){return b(`${y}/api/question/generateSupplementaryQuestion`,e,te)}generateNextCorrectAnswer(e){return f(`${y}/api/question/generateNextCorrectAnswer?questionId=${e}`,D)}addQuestionAnswer(e){return b(`${y}/api/question/addQuestionAnswer`,e,D)}addSupplementaryQuestionAnswer(e){return b(`${y}/api/question/addSupplementaryQuestionAnswer`,e,ee)}};ne=function(e,t,n,s){var r,i=arguments.length,a=i<3?t:null===s?s=Object.getOwnPropertyDescriptor(t,n):s;if("object"==typeof Reflect&&"function"==typeof Reflect.decorate)a=Reflect.decorate(e,t,n,s);else for(var o=e.length-1;o>=0;o--)(r=e[o])&&(a=(i<3?r(a):i>3?r(t,n,a):r(t,n))||a);return i>3&&a&&Object.defineProperty(t,n,a),a}([(0,i._G)()],ne);var se=function(e,t,n,s){var r,i=arguments.length,a=i<3?t:null===s?s=Object.getOwnPropertyDescriptor(t,n):s;if("object"==typeof Reflect&&"function"==typeof Reflect.decorate)a=Reflect.decorate(e,t,n,s);else for(var o=e.length-1;o>=0;o--)(r=e[o])&&(a=(i<3?r(a):i>3?r(t,n,a):r(t,n))||a);return i>3&&a&&Object.defineProperty(t,n,a),a},re=function(e,t){if("object"==typeof Reflect&&"function"==typeof Reflect.metadata)return Reflect.metadata(e,t)};let ie=class{constructor(e){Object.defineProperty(this,"questionController",{enumerable:!0,configurable:!0,writable:!0,value:e}),Object.defineProperty(this,"isFeedbackVisible",{enumerable:!0,configurable:!0,writable:!0,value:!0}),Object.defineProperty(this,"isQuestionFreezed",{enumerable:!0,configurable:!0,writable:!0,value:!1}),Object.defineProperty(this,"feedback",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"question",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"lastAnswer",{enumerable:!0,configurable:!0,writable:!0,value:[]}),Object.defineProperty(this,"answersHistory",{enumerable:!0,configurable:!0,writable:!0,value:[]}),Object.defineProperty(this,"supplementaryQuestion",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"questionState",{enumerable:!0,configurable:!0,writable:!0,value:"INITIAL"}),Object.defineProperty(this,"storeState",{enumerable:!0,configurable:!0,writable:!0,value:{tag:"VALID"}}),Object.defineProperty(this,"onQuestionLoaded",{enumerable:!0,configurable:!0,writable:!0,value:e=>{var t,n;e.options.requireContext&&[...e.text.matchAll(/(\<\w.*?\sid\s*?\=([\'\"]))\s*(answer_(\d+?))\2(.*?\>)/gim)].forEach(((t,n)=>{e.text=e.text.replace(t[0],`${t[1]}question_${e.questionId}_${t[3]}_${n}${t[2]} data-answer-id='${t[4]}' ${t[5]}`)})),this.question=e,this.supplementaryQuestion=new j(this.questionController,e.questionId),this.feedback=null!==(t=e.feedback)&&void 0!==t?t:void 0,this.isFeedbackVisible=!0,this.answersHistory=[],this.lastAnswer=null!==(n=e.responses)&&void 0!==n?n:[],e.feedback&&0===e.feedback.stepsLeft&&this.setQuestionState("COMPLETED")}}),Object.defineProperty(this,"setQuestionState",{enumerable:!0,configurable:!0,writable:!0,value:e=>{this.questionState!==e&&(this.questionState=e)}}),Object.defineProperty(this,"setValidStoreState",{enumerable:!0,configurable:!0,writable:!0,value:()=>{"VALID"!==this.storeState.tag&&(this.storeState={tag:"VALID"})}}),Object.defineProperty(this,"setErrorStoreState",{enumerable:!0,configurable:!0,writable:!0,value:e=>{this.storeState={tag:"ERROR",error:e}}}),Object.defineProperty(this,"loadQuestion",{enumerable:!0,configurable:!0,writable:!0,value:(0,O.L3)((function*(e){this.setValidStoreState(),this.setQuestionState("LOADING");const t=yield this.questionController.getQuestion(e);this.setQuestionState("LOADED"),g.isLeft(t)?this.setErrorStoreState(t.left):this.onQuestionLoaded(t.right)}))}),Object.defineProperty(this,"generateQuestion",{enumerable:!0,configurable:!0,writable:!0,value:(0,O.L3)((function*(e){this.setValidStoreState(),this.setQuestionState("LOADING");const t=yield this.questionController.generateQuestionByAttempt(e);this.setQuestionState("LOADED"),g.isLeft(t)?this.setErrorStoreState(t.left):this.onQuestionLoaded(t.right)}))}),Object.defineProperty(this,"generateQuestionByMetadata",{enumerable:!0,configurable:!0,writable:!0,value:(0,O.L3)((function*(e){this.setValidStoreState(),this.setQuestionState("LOADING");const t=yield this.questionController.generateQuestionByMetadata(e);this.setQuestionState("LOADED"),g.isLeft(t)?this.setErrorStoreState(t.left):this.onQuestionLoaded(t.right)}))}),Object.defineProperty(this,"generateNextCorrectAnswer",{enumerable:!0,configurable:!0,writable:!0,value:(0,O.L3)((function*(){const{question:e}=this;if(!e)throw new Error("Current question not found");this.setValidStoreState(),this.setQuestionState("ANSWER_EVALUATING");const t=yield this.questionController.generateNextCorrectAnswer(e.questionId);this.setQuestionState("LOADED"),g.isLeft(t)?this.setErrorStoreState(t.left):this.onAnswerEvaluated(t.right)}))}),Object.defineProperty(this,"sendAnswersImpl",{enumerable:!0,configurable:!0,writable:!0,value:(0,O.L3)((function*(e,t){const n=(0,O.HO)({questionId:e,answers:(0,O.HO)([...t])});this.setValidStoreState(),this.setQuestionState("ANSWER_EVALUATING");const s=yield this.questionController.addQuestionAnswer(n);this.setQuestionState("LOADED"),g.isLeft(s)?this.setErrorStoreState(s.left):this.onAnswerEvaluated(s.right)}))}),Object.defineProperty(this,"sendAnswers",{enumerable:!0,configurable:!0,writable:!0,value:(0,O.L3)((function*(){const{question:e,lastAnswer:t}=this;e&&(yield this.sendAnswersImpl(e.questionId,(0,O.HO)(t)))}))}),Object.defineProperty(this,"onAnswersChanged",{enumerable:!0,configurable:!0,writable:!0,value:(0,O.L3)((function*(e,t=!0){if(this.answersHistory.push(e),t)try{yield this.sendAnswers()}catch(e){this.answersHistory.pop()}}))}),Object.defineProperty(this,"setFullAnswer",{enumerable:!0,configurable:!0,writable:!0,value:(0,O.L3)((function*(e,t=!0){if(!this.isAnswerChanged(e))return!1;const n=this.lastAnswer;if(this.lastAnswer=e,n.length>0&&this.answersHistory.push(n),!t)return!0;try{return yield this.sendAnswers(),!0}catch(e){return this.lastAnswer=n,n.length>0&&this.answersHistory.pop(),!1}}))}),Object.defineProperty(this,"isAnswerChanged",{enumerable:!0,configurable:!0,writable:!0,value:e=>{const{lastAnswer:t,question:n}=this;if(!n)throw new Error("no question");const s=t.map((e=>e.answer)),r=e.map((e=>e.answer));switch(n.type){case"ORDER":return r.length!==s.length||JSON.stringify(r)!==JSON.stringify(s);case"MATCHING":case"MULTI_CHOICE":case"SINGLE_CHOICE":return r.length!==s.length||JSON.stringify(r.sort())!==JSON.stringify(s.sort())}}}),(0,O.Gn)(this)}onAnswerEvaluated(e){this.feedback=e,this.isFeedbackVisible=!0,e&&e.correctAnswers&&(this.setFullAnswer(e.correctAnswers,!1),c(e.stepsLeft)||0!==e.stepsLeft||this.setQuestionState("COMPLETED"))}};var ae,oe;se([O.sH,re("design:type",Boolean)],ie.prototype,"isFeedbackVisible",void 0),se([O.sH,re("design:type",Boolean)],ie.prototype,"isQuestionFreezed",void 0),se([O.sH,re("design:type",Object)],ie.prototype,"feedback",void 0),se([O.sH,re("design:type",Object)],ie.prototype,"question",void 0),se([O.sH,re("design:type",Array)],ie.prototype,"lastAnswer",void 0),se([O.sH,re("design:type",Array)],ie.prototype,"answersHistory",void 0),se([O.sH,re("design:type",j)],ie.prototype,"supplementaryQuestion",void 0),se([O.sH,re("design:type",String)],ie.prototype,"questionState",void 0),se([O.sH,re("design:type",Object)],ie.prototype,"storeState",void 0),se([O.XI,re("design:type",Object)],ie.prototype,"setQuestionState",void 0),se([O.XI,re("design:type",Object)],ie.prototype,"setValidStoreState",void 0),se([O.XI,re("design:type",Object)],ie.prototype,"setErrorStoreState",void 0),ie=se([(0,i._G)(),(ae=0,oe=(0,i.WQ)(ne),function(e,t){oe(e,t,ae)}),re("design:paramtypes",[Object])],ie);var le=n(2635);const ce=a.KC([a.NW({kind:a.eu("AFTER_FIRST")}),a.NW({kind:a.eu("AFTER_LAST")}),a.NW({kind:a.eu("AFTER_EACH")}),a.NW({kind:a.eu("AFTER_SPECIFIC"),numbers:a.YO(a.ai)})]),ue=a.NW({id:a.ai,type:a.eu("yes-no"),text:a.Yj,policy:ce,required:a.zM,options:a.NW({yesText:a.Yj,yesValue:a.Yj,noText:a.Yj,noValue:a.Yj})},"YesNoSurveyQuestion"),de=a.NW({id:a.ai,type:a.eu("single-choice"),text:a.Yj,policy:ce,required:a.zM,options:a.YO(a.NW({id:a.Yj,text:a.Yj}))}),me=a.NW({id:a.ai,type:a.eu("open-ended"),text:a.Yj,policy:ce,required:a.zM}),pe=a.KC([ue,de,me]),ge=a.NW({surveyId:a.Yj,options:a.NW({}),questions:a.YO(pe)},"Survey"),he=a.NW({surveyQuestionId:a.ai,questionId:a.ai,answer:a.Yj});var ve=function(e,t,n,s){return new(n||(n=Promise))((function(r,i){function a(e){try{l(s.next(e))}catch(e){i(e)}}function o(e){try{l(s.throw(e))}catch(e){i(e)}}function l(e){var t;e.done?r(e.value):(t=e.value,t instanceof n?t:new n((function(e){e(t)}))).then(a,o)}l((s=s.apply(e,t||[])).next())}))};let fe=class{constructor(){Object.defineProperty(this,"surveyCache",{enumerable:!0,configurable:!0,writable:!0,value:{}})}getSurvey(e){return ve(this,void 0,void 0,(function*(){if(this.surveyCache[e])return g.right(this.surveyCache[e]);var t=yield f(`${y}/api/survey/${e}`,ge);return g.isRight(t)&&(this.surveyCache[e]=t.right),t}))}postSurveyAnswer(e,t,n){return ve(this,void 0,void 0,(function*(){return b(`${y}/api/survey`,{surveyQuestionId:e,questionId:t,answer:n})}))}getCurrentUserAttemptSurveyVotes(e,t){return ve(this,void 0,void 0,(function*(){return f(`${y}/api/survey/${encodeURIComponent(e)}/user-votes?attemptId=${t}`,a.YO(he))}))}};fe=function(e,t,n,s){var r,i=arguments.length,a=i<3?t:null===s?s=Object.getOwnPropertyDescriptor(t,n):s;if("object"==typeof Reflect&&"function"==typeof Reflect.decorate)a=Reflect.decorate(e,t,n,s);else for(var o=e.length-1;o>=0;o--)(r=e[o])&&(a=(i<3?r(a):i>3?r(t,n,a):r(t,n))||a);return i>3&&a&&Object.defineProperty(t,n,a),a}([(0,i._G)()],fe);var be=n(2514),Ee=n(7042),we=n(3702);function ye(e){return new URLSearchParams(window.location.search).get(e)}const xe=a.ZE({EN:null,RU:null,PL:null}),Ce=a.NW({id:a.ai,displayName:a.Yj,email:a.KC([a.Yj,a.ch]),language:xe,roles:a.YO(a.Yj)},"UserInfo");let Se=class{getCurrentUser(){return f(`${y}/api/users/whoami`,Ce)}};Se=function(e,t,n,s){var r,i=arguments.length,a=i<3?t:null===s?s=Object.getOwnPropertyDescriptor(t,n):s;if("object"==typeof Reflect&&"function"==typeof Reflect.decorate)a=Reflect.decorate(e,t,n,s);else for(var o=e.length-1;o>=0;o--)(r=e[o])&&(a=(i<3?r(a):i>3?r(t,n,a):r(t,n))||a);return i>3&&a&&Object.defineProperty(t,n,a),a}([(0,i._G)()],Se);var _e=function(e,t,n,s){var r,i=arguments.length,a=i<3?t:null===s?s=Object.getOwnPropertyDescriptor(t,n):s;if("object"==typeof Reflect&&"function"==typeof Reflect.decorate)a=Reflect.decorate(e,t,n,s);else for(var o=e.length-1;o>=0;o--)(r=e[o])&&(a=(i<3?r(a):i>3?r(t,n,a):r(t,n))||a);return i>3&&a&&Object.defineProperty(t,n,a),a},Ne=function(e,t){if("object"==typeof Reflect&&"function"==typeof Reflect.metadata)return Reflect.metadata(e,t)},Ae=function(e,t){return function(n,s){t(n,s,e)}},Oe=function(e,t,n,s){return new(n||(n=Promise))((function(r,i){function a(e){try{l(s.next(e))}catch(e){i(e)}}function o(e){try{l(s.throw(e))}catch(e){i(e)}}function l(e){var t;e.done?r(e.value):(t=e.value,t instanceof n?t:new n((function(e){e(t)}))).then(a,o)}l((s=s.apply(e,t||[])).next())}))};let Ie=class{constructor(e,t,n,s){Object.defineProperty(this,"exerciseController",{enumerable:!0,configurable:!0,writable:!0,value:e}),Object.defineProperty(this,"userController",{enumerable:!0,configurable:!0,writable:!0,value:t}),Object.defineProperty(this,"surveyController",{enumerable:!0,configurable:!0,writable:!0,value:n}),Object.defineProperty(this,"isSessionLoading",{enumerable:!0,configurable:!0,writable:!0,value:!1}),Object.defineProperty(this,"user",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"exerciseId",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"exercise",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"currentAttemptId",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"currentAttempt",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"currentQuestion",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"exerciseState",{enumerable:!0,configurable:!0,writable:!0,value:"INITIAL"}),Object.defineProperty(this,"storeState",{enumerable:!0,configurable:!0,writable:!0,value:{tag:"VALID"}}),Object.defineProperty(this,"survey",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"isDebug",{enumerable:!0,configurable:!0,writable:!0,value:!1}),Object.defineProperty(this,"registerOnStrategyDecisionChangedAction",{enumerable:!0,configurable:!0,writable:!0,value:()=>{(0,O.fm)((()=>{var e;"FINISH"===(null===(e=this.currentQuestion.feedback)||void 0===e?void 0:e.strategyDecision)&&"COMPLETED"!==this.exerciseState&&this.setExerciseState("COMPLETED")}))}}),Object.defineProperty(this,"forceSetValidState",{enumerable:!0,configurable:!0,writable:!0,value:()=>{"VALID"!==this.storeState.tag&&(this.storeState={tag:"VALID"})}}),Object.defineProperty(this,"setExerciseState",{enumerable:!0,configurable:!0,writable:!0,value:e=>{this.exerciseState!==e&&(this.exerciseState=e)}}),Object.defineProperty(this,"setSurveyAnswers",{enumerable:!0,configurable:!0,writable:!0,value:(e,t)=>{this.survey&&(0,O.h5)((()=>{this.survey.questions[e].status="COMPLETED",this.survey.questions[e].results=t}))}}),Object.defineProperty(this,"loadSessionInfo",{enumerable:!0,configurable:!0,writable:!0,value:()=>Oe(this,void 0,void 0,(function*(){if(this.exercise)throw new Error("exerciseInfo loaded");if(this.isSessionLoading)return;(0,O.h5)((()=>{this.forceSetValidState(),this.isSessionLoading=!0}));const[e,t]=yield Promise.all([this.userController.getCurrentUser(),this.exerciseController.getExerciseShortInfo(this.exerciseId)]);g.isRight(e)&&g.isRight(t)?(0,O.h5)((()=>{this.isSessionLoading=!1,this.onSessionLoaded(e.right,t.right)})):g.isLeft(e)&&g.isLeft(t)&&(0,O.h5)((()=>{var n;this.isSessionLoading=!1,this.storeState={tag:"ERROR",error:null!==(n=e.left)&&void 0!==n?n:t.left}}))}))}),Object.defineProperty(this,"loadExerciseAttempt",{enumerable:!0,configurable:!0,writable:!0,value:(0,O.L3)((function*(e){const{exercise:t}=this;if(!t)throw new Error("exerciseInfo is not defined");this.forceSetValidState(),t.id;const n=yield this.exerciseController.getExerciseAttempt(e);if(g.isLeft(n))return void(this.storeState={tag:"ERROR",error:n.left});const s=n.right;return!!s&&(this.currentAttempt=s,yield this.onAttemptLoaded(),!0)}))}),Object.defineProperty(this,"loadExistingExerciseAttempt",{enumerable:!0,configurable:!0,writable:!0,value:(0,O.L3)((function*(){const{exercise:e}=this;if(!e)throw new Error("exercise is not defined");this.forceSetValidState();const t=e.id,n=yield this.exerciseController.getExistingExerciseAttempt(t);if(g.isLeft(n))return void(this.storeState={tag:"ERROR",error:n.left});const s=n.right;return!!s&&(this.currentAttempt=s,yield this.onAttemptLoaded(),!0)}))}),Object.defineProperty(this,"onAttemptLoaded",{enumerable:!0,configurable:!0,writable:!0,value:()=>Oe(this,void 0,void 0,(function*(){yield this.loadSurvey()}))}),Object.defineProperty(this,"createExerciseAttempt",{enumerable:!0,configurable:!0,writable:!0,value:(0,O.L3)((function*(){const{exercise:e}=this;if(!e)throw new Error("exercise is not defined");this.forceSetValidState();const t=e.id,n=yield this.exerciseController.createExerciseAttempt(+t);g.isLeft(n)?this.storeState={tag:"ERROR",error:n.left}:(this.currentAttempt=n.right,yield this.onAttemptLoaded())}))}),Object.defineProperty(this,"createDebugExerciseAttempt",{enumerable:!0,configurable:!0,writable:!0,value:(0,O.L3)((function*(){const{exercise:e}=this;if(!e)throw new Error("exercise is not defined");this.forceSetValidState();const t=e.id,n=yield this.exerciseController.createDebugExerciseAttempt(+t);g.isLeft(n)?this.storeState={tag:"ERROR",error:n.left}:(this.currentAttempt=n.right,yield this.onAttemptLoaded())}))}),Object.defineProperty(this,"generateQuestion",{enumerable:!0,configurable:!0,writable:!0,value:(0,O.L3)((function*(){var e,t;const{exercise:n,currentAttempt:s}=this;if(!n||!s)throw new Error("Session is not defined");this.forceSetValidState(),yield this.currentQuestion.generateQuestion(s.attemptId),s.questionIds.push(null!==(t=null===(e=this.currentQuestion.question)||void 0===e?void 0:e.questionId)&&void 0!==t?t:-1)}))}),Object.defineProperty(this,"changeLanguage",{enumerable:!0,configurable:!0,writable:!0,value:e=>{this.user&&this.user.language!==e&&(this.user.language=e,le.Ay.changeLanguage(e))}}),Object.defineProperty(this,"loadSurvey",{enumerable:!0,configurable:!0,writable:!0,value:()=>Oe(this,void 0,void 0,(function*(){var e;if(this.survey||!this.currentAttempt||!this.exercise)return;if(!(null===(e=this.exercise.options.surveyOptions)||void 0===e?void 0:e.enabled)||0===this.exercise.options.surveyOptions.surveyId.length)return;const t=this.exercise.options.surveyOptions.surveyId,n=this.currentAttempt.attemptId,[s,r]=yield Promise.all([this.surveyController.getSurvey(t),this.surveyController.getCurrentUserAttemptSurveyVotes(t,n)]);(0,O.h5)((()=>{if(g.isRight(s)&&g.isRight(r)){const e=function(e,t){const n=new Map;return e.forEach((e=>{const t=e.questionId,s=n.get(t);s?s.push(e):n.set(t,[e])})),n}(r.right);this.survey={survey:s.right,questions:[...e.keys()].map((t=>{var n,s,r,i;return{questionId:t,status:"COMPLETED",questions:null!==(s=null===(n=e.get(t))||void 0===n?void 0:n.map((e=>e.surveyQuestionId)))&&void 0!==s?s:[],results:null!==(i=null===(r=e.get(t))||void 0===r?void 0:r.reduce(((e,t)=>(e[t.surveyQuestionId]=t.answer,e)),{}))&&void 0!==i?i:{}}})).reduce(((e,t)=>(e[t.questionId]=t,e)),{})}}}))}))}),Object.defineProperty(this,"ensureQuestionSurveyExists",{enumerable:!0,configurable:!0,writable:!0,value:e=>{var t,n,s;if(null===(t=this.survey)||void 0===t?void 0:t.questions[e])return null===(n=this.survey)||void 0===n?void 0:n.questions[e].questions;const r=[],i=this.currentAttempt.questionIds.findIndex((e=>{var t;return e===(null===(t=this.currentQuestion.question)||void 0===t?void 0:t.questionId)}));for(let e of(null===(s=this.survey)||void 0===s?void 0:s.survey.questions)||[]){const t=e.policy;("AFTER_EACH"===t.kind||"AFTER_FIRST"===t.kind&&0===i||"AFTER_LAST"===t.kind&&"COMPLETED"===this.exerciseState||"AFTER_SPECIFIC"===t.kind&&t.numbers.includes(i+1))&&r.push(e)}console.log("Selected questions"),console.log((0,O.HO)(r));var a={questionId:e,status:"ACTIVE",questions:r.map((e=>e.id)),results:{}};return(0,O.h5)((()=>{this.survey.questions[e]=a})),r.map((e=>e.id))}}),this.isDebug=null!==ye("debug"),this.currentQuestion=s;const r=ye("exerciseId");null===r&&(this.exerciseState="LAUNCH_ERROR",this.storeState={tag:"ERROR",error:{message:"Invalid exercise id"}}),this.exerciseId=null!==r?+r:-1;const i=ye("attemptId");null!==i&&(this.currentAttemptId=+i),(0,O.Gn)(this),this.registerOnStrategyDecisionChangedAction()}onSessionLoaded(e,t){(0,O.h5)((()=>{this.user=e,this.exercise=t,e.language!==le.Ay.language&&le.Ay.changeLanguage(e.language)}))}};_e([O.sH,Ne("design:type",Boolean)],Ie.prototype,"isSessionLoading",void 0),_e([O.sH,Ne("design:type",Object)],Ie.prototype,"user",void 0),_e([O.sH,Ne("design:type",Number)],Ie.prototype,"exerciseId",void 0),_e([O.sH,Ne("design:type",Object)],Ie.prototype,"exercise",void 0),_e([O.sH,Ne("design:type",Number)],Ie.prototype,"currentAttemptId",void 0),_e([O.sH,Ne("design:type",Object)],Ie.prototype,"currentAttempt",void 0),_e([O.sH,Ne("design:type",ie)],Ie.prototype,"currentQuestion",void 0),_e([O.sH,Ne("design:type",String)],Ie.prototype,"exerciseState",void 0),_e([O.sH,Ne("design:type",Object)],Ie.prototype,"storeState",void 0),_e([O.sH,Ne("design:type",Object)],Ie.prototype,"survey",void 0),_e([O.sH,Ne("design:type",Object)],Ie.prototype,"isDebug",void 0),_e([O.XI,Ne("design:type",Object)],Ie.prototype,"forceSetValidState",void 0),_e([O.XI,Ne("design:type",Object)],Ie.prototype,"setExerciseState",void 0),_e([O.XI,Ne("design:type",Object)],Ie.prototype,"setSurveyAnswers",void 0),_e([O.XI,Ne("design:type",Object)],Ie.prototype,"changeLanguage",void 0),_e([O.XI,Ne("design:type",Object)],Ie.prototype,"loadSurvey",void 0),_e([O.XI,Ne("design:type",Object)],Ie.prototype,"ensureQuestionSurveyExists",void 0),Ie=_e([(0,i._G)(),Ae(0,(0,i.WQ)(S)),Ae(1,(0,i.WQ)(Se)),Ae(2,(0,i.WQ)(fe)),Ae(3,(0,i.WQ)(ie)),Ne("design:paramtypes",[Object,Object,fe,ie])],Ie);var qe=n(2389);const ke=a.NW({id:a.ai,name:a.Yj}),Le=a.NW({name:a.Yj,kind:a.ZE({FORBIDDEN:null,PERMITTED:null,TARGETED:null})}),je=a.NW({name:a.Yj,kind:a.ZE({FORBIDDEN:null,PERMITTED:null,TARGETED:null})}),Qe=a.NW({name:a.Yj,kind:a.ZE({FORBIDDEN:null,PERMITTED:null,TARGETED:null})}),Te=a.NW({numberOfQuestions:a.ai,complexity:a.ai,concepts:a.YO(Le),laws:a.YO(je),skills:a.YO(Qe)}),Pe=a.NW({id:a.ai,name:a.Yj,domainId:a.Yj,strategyId:a.Yj,backendId:a.Yj,stages:function(e,t=`NonEmptyArray<${e.name}>`){const n=a.YO(e);return new a.ZU(t,(e=>n.is(e)&&(0,Ee.isNonEmpty)(e)),((e,t)=>(0,I.pipe)(n.validate(e,t),(0,g.chain)((n=>{const s=(0,be.fromArray)(n);return(0,we.isNone)(s)?a.V2(e,t):a.kX(s.value)})))),(0,be.map)(e.encode))}(Te),tags:a.YO(a.Yj),options:x}),De=a.FX("DomainSkill",(()=>a.NW({name:a.Yj,displayName:a.Yj,childs:a.YO(De)}))),Re=a.FX("DomainLaw",(()=>a.NW({name:a.Yj,displayName:a.Yj,bitflags:a.ai,childs:a.YO(Re)})));var Me;!function(e){e[e.VisibleToTeacher=1]="VisibleToTeacher",e[e.TargetEnabled=2]="TargetEnabled"}(Me||(Me={}));const $e=a.FX("DomainConcept",(()=>a.NW({name:a.Yj,displayName:a.Yj,bitflags:a.ai,childs:a.YO($e)}))),He=a.NW({id:a.Yj,displayName:a.Yj,description:a.KC([a.Yj,a.ch]),laws:a.YO(Re),skills:a.YO(De),concepts:a.YO($e),tags:a.YO(a.Yj)}),Fe=a.NW({id:a.Yj,displayName:a.Yj,description:a.KC([a.Yj,a.ch]),options:a.NW({multiStagesEnabled:a.zM})}),Ve=a.NW({count:a.ai,topRatedCount:a.ai,questions:a.YO(a.NW({metadataId:a.ai,name:a.Yj}))});let Ge=class{getAllExercises(){return f(`${y}/api/exercise/all`,a.YO(ke))}getExercise(e){return f(`${y}/api/exercise?id=${encodeURIComponent(e)}`,Pe)}saveExercise(e){return b(`${y}/api/exercise`,(0,O.HO)(e))}createExercise(e,t,n){return function(e,t,n,s){return h(this,void 0,void 0,(function*(){const s=Object.assign(Object.assign({},v),{method:"PUT",body:JSON.stringify(t),signal:undefined});return yield E(e,s,n)}))}(`${y}/api/exercise`,{name:e,domainId:t,strategyId:n},a.ai)}deleteExercise(e){return function(e,t,n){return h(this,void 0,void 0,(function*(){const t=Object.assign(Object.assign({},v),{method:"DELETE",signal:undefined});return yield E(e,t,undefined)}))}(`${y}/api/exercise?id=${encodeURIComponent(e)}`)}getStrategies(){return f(`${y}/api/refTables/strategies`,a.YO(Fe))}getBackends(){return f(`${y}/api/refTables/backends`,a.YO(a.Yj))}getDomains(){return f(`${y}/api/refTables/domains`,a.YO(He))}getDomainLaws(e){return f(`${y}/api/refTables/domainLaws?domaindId=${encodeURIComponent(e)}`,a.YO(a.Yj))}getDomainConcepts(e){return f(`${y}/api/refTables/domainConcepts?domaindId=${encodeURIComponent(e)}`,a.YO(a.Yj))}search(e,t,n,s,r,i,a,o){return b(`${y}/api/question-bank/search`,{domainId:e,tags:r,concepts:t,laws:n,skills:s,complexity:i,limit:a},Ve,o)}};Ge=function(e,t,n,s){var r,i=arguments.length,a=i<3?t:null===s?s=Object.getOwnPropertyDescriptor(t,n):s;if("object"==typeof Reflect&&"function"==typeof Reflect.decorate)a=Reflect.decorate(e,t,n,s);else for(var o=e.length-1;o>=0;o--)(r=e[o])&&(a=(i<3?r(a):i>3?r(t,n,a):r(t,n))||a);return i>3&&a&&Object.defineProperty(t,n,a),a}([(0,i._G)()],Ge);const Be=()=>{var e;return null!==(null!==(e=new URLSearchParams(window.location.search).get("sandbox"))&&void 0!==e?e:null)};i.kL.register(S,{useFactory:()=>Be()?new N:new S}),i.kL.register(ne,{useFactory:()=>Be()?new N:new ne}),i.kL.register(Se,{useFactory:()=>Be()?new N:new Se}),i.kL.register(ie,ie),i.kL.registerSingleton(Ie),i.kL.registerSingleton(fe),i.kL.registerSingleton(Ge),le.Ay.use(qe.r9).init({resources:{EN:{translation:{question_header:"Question #{{questionNumber}}",language_header:"Language",signedin_as_header:"Signed in as",nextCorrectAnswerBtn:"I'm confused, tell me the next correct step",generateNextQuestion_nextQuestion:"Next question",generateNextQuestion_warning:"Warning",generateNextQuestion_continueAttempt:"Continue attempt",generateNextQuestion_modalMessage1:"It is recommended to move on to the next question only after solving all the previous ones.",generateNextQuestion_modalMessage2:"Are you sure you want to move on to the next question?",grade_feeback:"Grade",correctsteps_feeback:"Correct steps",stepswitherrors_feeback:"Steps with errors",stepsleft_feeback:"Steps left",issolved_feeback:"Solved",foundExisitingAttempt_title:"Found existing attempt",foundExisitingAttempt_descr:"Would you like to continue the existing attempt or start a new one",foundExisitingAttempt_continueattempt:"Continue",foundExisitingAttempt_newattempt:"New",exercise_completed:"Exercise completed",exercise_supquestion_gotit:"Got it",exercise_supquestion_details:"More details",exercise_supquestion_send_answer:"Send answer",exercise_supquestion_next_question:"Next question",exercisesettings_title:"Exercise settings",exercisesettings_name:"Name",exercisesettings_domain:"Domain",exercisesettings_strategy:"Strategy",exercisesettings_qcomplexity:"Question difficulty",exercisesettings_answlen:"Answer length",exercisesettings_qopt:"Options",exercisesettings_qopt_forceAttCreation:"Always create a new attempt",exercisesettings_qopt_genCorAnsw:"Allow 'generate correct answer' button",exercisesettings_qopt_forceShowGenNextQ:"Always show 'generate new question' button",exercisesettings_qopt_supQ:"Allow supplementary questions",exercisesettings_qopt_preferDTsup:"Prefer Decision-Tree-based approach to supplementary question generation",exercisesettings_survey:"Survey",exercisesettings_tags:"Tags",exercisesettings_commonConcepts:"Common concepts",exercisesettings_commonLaws:"Common laws",exercisesettings_commonSkills:"Common skills",exercisesettings_stages:"Stage",exercisesettings_stageN:"Stage #{{stageNumber}}",exercisesettings_stageN_qnumber:"Number of questions",exercisesettings_stageN_concepts:"Concepts",exercisesettings_stageN_laws:"Laws",exercisesettings_stageN_skills:"Skills",exercisesettings_stageN_matchedQuestionExamples:"Question examples",exercisesettings_addStage:"Add stage",exercisesettings_removeStage:"Remove stage",exercisesettings_save:"Save",exercisesettings_saveNopen:"Save & Open",exercisesettings_open:"Open",exercisesettings_genDebugAtt:"Generate debug attempt",exercisesettings_optDenied:"Denied",exercisesettings_optAllowed:"Allowed",exercisesettings_optTarget:"Target",exercisesettings_questionsInBank:"Questions in bank",exercisesettings_noQuestionsFound:"No suitable questions found",survey_sendresults:"Send survey results"}},RU:{translation:{question_header:"Вопрос #{{questionNumber}}",language_header:"Язык",signedin_as_header:"Пользователь",nextCorrectAnswerBtn:"Я в замешательстве, подскажи следующий корректный шаг",generateNextQuestion_nextQuestion:"Следующий вопрос",generateNextQuestion_warning:"Предупреждение",generateNextQuestion_continueAttempt:"Продолжить попытку",generateNextQuestion_modalMessage1:"Рекомендуется переходить к следующему вопросу только после решения всех предыдущих.",generateNextQuestion_modalMessage2:"Вы действительно хотите перейти к следующему вопросу?",grade_feeback:"Оценка",correctsteps_feeback:"Правильных шагов",stepswitherrors_feeback:"Шагов с ошибками",stepsleft_feeback:"Шагов осталось",issolved_feeback:"Задача решена",foundExisitingAttempt_title:"Найдена неоконченная попытка",foundExisitingAttempt_descr:"Вы хотите продолжить существующую попытку или начать новую",foundExisitingAttempt_continueattempt:"Продолжить",foundExisitingAttempt_newattempt:"Новая",exercise_completed:"Упражнение завершено",exercise_supquestion_gotit:"Понятно",exercise_supquestion_details:"Разобраться подробнее",exercise_supquestion_send_answer:"Отправить ответ",exercise_supquestion_next_question:"Следующий вопрос",exercisesettings_title:"Настройка упражнений",exercisesettings_name:"Название",exercisesettings_domain:"Домен",exercisesettings_strategy:"Стратегия",exercisesettings_qcomplexity:"Сложность вопросов",exercisesettings_answlen:"Длина ответа",exercisesettings_qopt:"Опции",exercisesettings_qopt_forceAttCreation:"Всегда создавать новую попытку",exercisesettings_qopt_genCorAnsw:"Разрешить подсказку следующего шага",exercisesettings_qopt_forceShowGenNextQ:"Разрешить неоконченные вопросы",exercisesettings_qopt_supQ:"Разрешить вспомогательные вопросы",exercisesettings_qopt_preferDTsup:"Предпочитать генерацию вспомогательных вопросов по дереву рассуждений",exercisesettings_survey:"Опрос",exercisesettings_tags:"Теги",exercisesettings_commonConcepts:"Общие концепты",exercisesettings_commonLaws:"Общие законы",exercisesettings_commonSkills:"Общие умения",exercisesettings_stages:"Стадии упражнения",exercisesettings_stageN:"Стадия #{{stageNumber}}",exercisesettings_stageN_qnumber:"Количество вопросов",exercisesettings_stageN_concepts:"Концепты",exercisesettings_stageN_laws:"Законы",exercisesettings_stageN_skills:"Умения",exercisesettings_stageN_matchedQuestionExamples:"Примеры вопросов",exercisesettings_addStage:"Добавить стадию",exercisesettings_removeStage:"Удалить стадию",exercisesettings_save:"Сохранить",exercisesettings_saveNopen:"Сохранить & Открыть",exercisesettings_open:"Открыть",exercisesettings_genDebugAtt:"Создать отладочную попытку",exercisesettings_optDenied:"Запрет",exercisesettings_optAllowed:"Разреш.",exercisesettings_optTarget:"Цель",exercisesettings_questionsInBank:"Вопросов в банке задач",exercisesettings_noQuestionsFound:"Подходящих вопросов не найдено",survey_sendresults:"Отправить результаты опроса"}},PL:{translation:{question_header:"Pytanie #{{questionNumber}}",language_header:"Język",signedin_as_header:"Zalogowany jako",nextCorrectAnswerBtn:"Nie wiem co robić dalej, podpowiedz mi następny poprawny krok",generateNextQuestion_nextQuestion:"Następne pytanie",generateNextQuestion_warning:"Ostrzeżenie",generateNextQuestion_continueAttempt:"Kontynuuj podejście",generateNextQuestion_modalMessage1:"Zaleca się przejście do następnego pytania dopiero po rozwiązaniu wszystkich poprzednich.",generateNextQuestion_modalMessage2:"Czy na pewno chcesz przejść do następnego pytania?",grade_feeback:"Ocena",correctsteps_feeback:"Poprawne kroki",stepswitherrors_feeback:"Kroki z błędami",stepsleft_feeback:"Pozostałe kroki",issolved_feeback:"Rozwiązane",foundExisitingAttempt_title:"Znaleziono już istniejące podejście",foundExisitingAttempt_descr:"Czy chcesz kontynuować dotychczasowe podejście, czy rozpocząć nowe",foundExisitingAttempt_continueattempt:"Kontynuuj",foundExisitingAttempt_newattempt:"Nowe",exercise_completed:"Ćwiczenie zakończone",exercise_supquestion_gotit:"Oczywiście!",exercise_supquestion_details:"Zobacz szczegóły",exercise_supquestion_send_answer:"Wyślij odpowiedź",exercise_supquestion_next_question:"Następne pytanie",exercisesettings_title:"Exercise settings",exercisesettings_name:"Name",exercisesettings_domain:"Domain",exercisesettings_strategy:"Strategy",exercisesettings_qcomplexity:"Question difficulty",exercisesettings_answlen:"Answer length",exercisesettings_qopt:"Options",exercisesettings_qopt_forceAttCreation:"Always create a new attempt",exercisesettings_qopt_genCorAnsw:"Allow 'generate correct answer' button",exercisesettings_qopt_forceShowGenNextQ:"Always show 'generate new question' button",exercisesettings_qopt_supQ:"Allow supplementary questions",exercisesettings_qopt_preferDTsup:"Prefer Decision-Tree-based approach to supplementary question generation",exercisesettings_survey:"Survey",exercisesettings_tags:"Tags",exercisesettings_commonConcepts:"Common concepts",exercisesettings_commonLaws:"Common laws",exercisesettings_stages:"Stage",exercisesettings_stageN:"Stage #{{stageNumber}}",exercisesettings_stageN_qnumber:"Number of questions",exercisesettings_stageN_concepts:"Concepts",exercisesettings_stageN_laws:"Laws",exercisesettings_stageN_matchedQuestionExamples:"Question examples",exercisesettings_addStage:"Add stage",exercisesettings_removeStage:"Remove stage",exercisesettings_save:"Save",exercisesettings_saveNopen:"Save & Open",exercisesettings_open:"Open",exercisesettings_genDebugAtt:"Generate debug attempt",exercisesettings_optDenied:"Denied",exercisesettings_optAllowed:"Allowed",exercisesettings_optTarget:"Target",exercisesettings_questionsInBank:"Questions in bank",exercisesettings_noQuestionsFound:"No suitable questions found",survey_sendresults:"Send survey results"}}},lng:"EN",interpolation:{escapeValue:!1}});var Ye=n(4976),ze=n(7767),We=n(4015),Ue=n(5615),Xe=n(3994);const Ke=(0,We.PA)((e=>{const{isVisible:t,children:n}=e;return t?r.createElement(r.Fragment,null,n):null})),Ze=e=>{const{title:t,primaryBtnTitle:n,handlePrimaryBtnClicked:s,secondaryBtnTitle:i,handleSecondaryBtnClicked:a,children:o,closeButton:l,show:u,handleClose:d,type:m,size:p}=e;return r.createElement(Ke,{isVisible:null==u||u},r.createElement(Je,{type:null!=m?m:"MODAL",show:null==u||u,onHide:null!=d?d:void 0,size:p},r.createElement(Ke,{isVisible:!c(t)&&t.length>0},r.createElement(Xe.A.Header,{closeButton:null!=l?l:void 0,placeholder:null},r.createElement(Xe.A.Title,null,t))),r.createElement(Xe.A.Body,null,o),i||n?r.createElement(Xe.A.Footer,null,i&&r.createElement(Ue.A,{variant:"secondary",onClick:null!=a?a:void 0},i),n&&r.createElement(Ue.A,{variant:"primary",onClick:null!=s?s:void 0},n)):null))},Je=e=>{const{type:t,show:n,onHide:s,children:i,size:a}=e;return"DIALOG"===t?r.createElement(Xe.A.Dialog,{size:a},i):r.createElement(Xe.A,{size:a,show:n,onHide:s},i)};var et=function(e,t,n,s){return new(n||(n=Promise))((function(r,i){function a(e){try{l(s.next(e))}catch(e){i(e)}}function o(e){try{l(s.throw(e))}catch(e){i(e)}}function l(e){var t;e.done?r(e.value):(t=e.value,t instanceof n?t:new n((function(e){e(t)}))).then(a,o)}l((s=s.apply(e,t||[])).next())}))};const tt=(0,We.PA)((()=>{const[e]=(0,r.useState)((()=>i.kL.resolve(Ie))),{t}=(0,qe.Bd)(),[n,s]=(0,r.useState)(!1),{exercise:a,currentAttempt:o}=e,{question:l}=e.currentQuestion,c="ANSWER_EVALUATING"===e.currentQuestion.questionState,u="LOADING"===e.currentQuestion.questionState;if(!l||!a||!o||u||c)return null;const d=()=>{s(!1)},m=()=>et(void 0,void 0,void 0,(function*(){s(!1);const{questionIds:t=[]}=o,n=t.indexOf(l.questionId);n===t.length-1?yield e.generateQuestion():yield e.currentQuestion.loadQuestion(t[n+1])}));return r.createElement(r.Fragment,null,r.createElement(Ue.A,{onClick:()=>et(void 0,void 0,void 0,(function*(){var t;const{questionIds:n=[]}=o,r=n.indexOf(l.questionId)===n.length-1;0!==(null===(t=e.currentQuestion.feedback)||void 0===t?void 0:t.stepsLeft)&&r?s(!0):yield m()})),variant:"primary"},t("generateNextQuestion_nextQuestion")),r.createElement(Ze,{show:n,title:t("generateNextQuestion_warning"),type:"MODAL",size:"lg",primaryBtnTitle:t("generateNextQuestion_continueAttempt"),handlePrimaryBtnClicked:d,secondaryBtnTitle:t("generateNextQuestion_nextQuestion"),handleSecondaryBtnClicked:m,closeButton:!1,handleClose:d},r.createElement("div",null,t("generateNextQuestion_modalMessage1")),r.createElement("div",null,t("generateNextQuestion_modalMessage2"))))}));var nt=n(1069);const st=(0,We.PA)((()=>{var e,t;const[n]=(0,r.useState)((()=>i.kL.resolve(Ie)));if(!n.currentQuestion.question||!n.currentAttempt)return null;const{questionId:s}=n.currentQuestion.question,a=n.currentAttempt.questionIds.map(((e,t)=>({id:e,number:t+1})));if(0===a.length)return null;const o=null!==(t=null===(e=a.find((e=>e.id===s)))||void 0===e?void 0:e.number)&&void 0!==t?t:-1,l=o-5<=0?"BEGIN":o+5>a.length?"END":"MIDDLE",c=Math.floor(2.5);return r.createElement("div",{className:"d-flex justify-content-center"},r.createElement(nt.A,{className:"p-3",style:{marginBottom:"0 !important"}},r.createElement(Ke,{isVisible:a.length<=10},a.map((e=>r.createElement(nt.A.Item,{key:e.number,active:s===e.id,onClick:()=>n.currentQuestion.loadQuestion(e.id)},e.number)))),r.createElement(Ke,{isVisible:a.length>10},r.createElement(nt.A.First,{disabled:s===a[0].id,onClick:()=>n.currentQuestion.loadQuestion(a[0].id)}),r.createElement(nt.A.Prev,{disabled:s===a[0].id,onClick:()=>n.currentQuestion.loadQuestion(a[a.findIndex((e=>s===e.id))-1].id)}),r.createElement(Ke,{isVisible:"BEGIN"===l},a.filter((e=>e.number<=7)).map((e=>r.createElement(nt.A.Item,{key:e.number,active:s===e.id,onClick:()=>n.currentQuestion.loadQuestion(e.id)},e.number))),r.createElement(nt.A.Ellipsis,{disabled:!0}),r.createElement(nt.A.Item,{key:a[a.length-1].number,active:!1,onClick:()=>n.currentQuestion.loadQuestion(a[a.length-1].id)},a[a.length-1].number)),r.createElement(Ke,{isVisible:"MIDDLE"===l},r.createElement(nt.A.Item,{key:a[0].number,active:!1,onClick:()=>n.currentQuestion.loadQuestion(a[0].id)},a[0].number),r.createElement(nt.A.Ellipsis,{disabled:!0}),a.filter((e=>e.number>=o-c&&e.number<=o+c)).map((e=>r.createElement(nt.A.Item,{key:e.number,active:s===e.id,onClick:()=>n.currentQuestion.loadQuestion(e.id)},e.number))),r.createElement(nt.A.Ellipsis,{disabled:!0}),r.createElement(nt.A.Item,{key:a[a.length-1].number,active:!1,onClick:()=>n.currentQuestion.loadQuestion(a[a.length-1].id)},a[a.length-1].number)),r.createElement(Ke,{isVisible:"END"===l},r.createElement(nt.A.Item,{key:a[0].number,active:!1,onClick:()=>n.currentQuestion.loadQuestion(a[0].id)},a[0].number),r.createElement(nt.A.Ellipsis,{disabled:!0}),a.filter((e=>e.number>a.length-7)).map((e=>r.createElement(nt.A.Item,{key:e.number,active:s===e.id,onClick:()=>n.currentQuestion.loadQuestion(e.id)},e.number)))),r.createElement(nt.A.Next,{disabled:s===a[a.length-1].id,onClick:()=>n.currentQuestion.loadQuestion(a[a.findIndex((e=>s===e.id))+1].id)}),r.createElement(nt.A.Last,{disabled:s===a[a.length-1].id,onClick:()=>n.currentQuestion.loadQuestion(a[a.length-1].id)}))))}));var rt=n(2406);const it=(0,We.PA)((e=>{const{text:t,pagination:n,languageHint:s,language:i,onLanguageClicked:a,userHint:o,user:l,onUserClicked:c,userHref:u}=e;return r.createElement(rt.A,{className:"px-0"},t&&r.createElement("h5",null,t)||null,r.createElement(rt.A.Collapse,{className:"justify-content-end"},n,r.createElement(rt.A.Text,{className:"px-2"},s,": ",r.createElement("a",{href:"#",onClick:null!=a?a:void 0},i)),r.createElement(rt.A.Toggle,null),r.createElement(rt.A.Text,{className:"px-2"},o,": ",r.createElement("a",{href:null!=u?u:"#",onClick:null!=c?c:void 0},l))))})),at=(0,We.PA)((()=>{const[e]=(0,r.useState)((()=>i.kL.resolve(Ie))),{t,i18n:n}=(0,qe.Bd)(),{currentAttempt:s,exercise:a,user:o,currentQuestion:l}=e;if(!s||!a||!o)return null;const c=s.questionIds.findIndex((e=>{var t;return(null===(t=l.question)||void 0===t?void 0:t.questionId)===e}));return r.createElement(it,{text:-1!==c?t("question_header",{questionNumber:c+1}):"",pagination:r.createElement(st,null),languageHint:t("language_header"),language:o.language,userHint:t("signedin_as_header"),user:o.displayName})}));var ot=n(778);const lt=(0,We.PA)((e=>{var t;const n=null!==(t=e.delay)&&void 0!==t?t:0,[s,i]=(0,r.useState)(0===n);return(0,r.useEffect)((()=>{n>0&&setTimeout((()=>!s&&i(!0)),n)})),s?r.createElement(ot.A,{style:Object.assign({},e.styleOverride),animation:"border",variant:"primary"}):null})),ct=(0,We.PA)((e=>{const{children:t,isLoading:n}=e;return n?r.createElement(lt,Object.assign({},e)):r.createElement(r.Fragment,null,t)})),ut=(0,We.PA)((({explicitQuestionStore:e})=>{var t;const[n]=(0,r.useState)((()=>null==e?i.kL.resolve(Ie):e)),{t:s}=(0,qe.Bd)();if(n instanceof Ie&&!(null===(t=n.exercise)||void 0===t?void 0:t.options.correctAnswerGenerationEnabled))return null;if(n instanceof Ie){const{exercise:e,currentAttempt:t,currentQuestion:s}=n,{question:r,feedback:i}=s,a="ANSWER_EVALUATING"===s.questionState,o="LOADING"===s.questionState;if(!r||!e||!t||a||o||0===(null==i?void 0:i.stepsLeft))return null}return r.createElement(Ue.A,{onClick:()=>{n instanceof Ie?n.currentQuestion.generateNextCorrectAnswer():n.generateNextCorrectAnswer()},variant:"primary"},s("nextCorrectAnswerBtn"))}));var dt=n(8079),mt=n(961),pt=n(4853),gt=n(3921),ht=function(e,t){var n={};for(var s in e)Object.prototype.hasOwnProperty.call(e,s)&&t.indexOf(s)<0&&(n[s]=e[s]);if(null!=e&&"function"==typeof Object.getOwnPropertySymbols){var r=0;for(s=Object.getOwnPropertySymbols(e);r<s.length;r++)t.indexOf(s[r])<0&&Object.prototype.propertyIsEnumerable.call(e,s[r])&&(n[s[r]]=e[s[r]])}return n};const vt=(0,We.PA)((e=>{const{question:t}=e,{options:n}=t;switch(!0){case"combobox"===n.displayMode&&!n.requireContext:return r.createElement(bt,Object.assign({},e));case"combobox"===n.displayMode&&n.requireContext:return r.createElement(Et,Object.assign({},e));case"dragNdrop"===n.displayMode:return r.createElement(ft,Object.assign({},e))}return r.createElement("div",null,"Not Implemented")})),ft=(0,We.PA)((e=>{const{question:t,getAnswers:n,onChanged:s}=e;if("dragNdrop"!==t.options.displayMode)return null;const{groups:i=[]}=t,{options:a}=t,o=a.dropzoneStyle&&JSON.parse(a.dropzoneStyle)||{},l=a.dropzoneStyle&&JSON.parse(a.draggableStyle)||{};return(0,r.useEffect)((()=>{document.querySelectorAll(`[id^="question_${t.questionId}_answer_"]`).forEach((e=>{e.classList.add("comp-ph-dropzone"),Object.keys(o).forEach((t=>e.style[t]=o[t])),e.innerHTML=`<div class="comp-ph-dropzone-placeholder">${a.dropzoneHtml}</div>`}));const e=new dt.gL(document.querySelectorAll(".comp-ph-droppable-container"),{draggable:".comp-ph-draggable",dropzone:".comp-ph-dropzone",plugins:[dt.ry.gU],mirror:{constrainDimensions:!0}});e.on("drag:over",(()=>console.log("is out"))),e.on("droppable:stop",(e=>{var r,i,o,l,c,u,d;const m=null===(l=null===(o=null===(i=null===(r=null==e?void 0:e.data)||void 0===r?void 0:r.dragEvent)||void 0===i?void 0:i.data)||void 0===o?void 0:o.source)||void 0===l?void 0:l.id,p=null===(u=null===(c=null==e?void 0:e.data)||void 0===c?void 0:c.dropzone)||void 0===u?void 0:u.id;if(m&&p){if(a.multipleSelectionEnabled){const e=`dragAnswerWrapper_${null!==(d=m.split("_")[1])&&void 0!==d?d:""}`,t=document.getElementById(e),n=document.getElementById(m);e!==p&&t&&n&&(t.innerHTML=n.outerHTML)}setTimeout((()=>{const e=[...document.querySelectorAll(`[id^="question_${t.questionId}_answer_"] > [id^="dragAnswer_"]`)].map((e=>{var n,s,r;return[+(null!==(s=null===(n=e.parentElement)||void 0===n?void 0:n.id.split(`question_${t.questionId}_answer_`)[1])&&void 0!==s?s:""),+(null!==(r=null==e?void 0:e.id.split("dragAnswer_")[1])&&void 0!==r?r:"")]})),r=n();s(e.map((e=>{var t,n;return{answer:e,isСreatedByUser:null===(n=null===(t=r.find((t=>t.answer[0]===e[0]&&t.answer[1]===e[1])))||void 0===t?void 0:t.isСreatedByUser)||void 0===n||n}})))}),10)}}))}),[t.questionId]),r.createElement("div",null,r.createElement("div",{className:"row"},r.createElement("div",{className:"col-md"},a.requireContext?r.createElement("p",{className:"comp-ph-droppable-container comp-ph-question-text",dangerouslySetInnerHTML:{__html:t.text}}):r.createElement("p",{className:"d-flex flex-column comp-ph-droppable-container comp-ph-question-text"},t.answers.map((e=>r.createElement("div",{className:"d-flex flex-row mb-3"},r.createElement("div",{className:"mr-2 mt-1"},r.createElement("div",{id:`question_${t.questionId}_answer_${e.id}`})),r.createElement("div",{dangerouslySetInnerHTML:{__html:e.text}})))))),r.createElement("div",{className:"col-md comp-ph-droppable-container d-flex justify-content-start align-items-start flex-column"},i.map((e=>r.createElement("div",{id:`dragAnswerWrapper_${e.id}`,className:"comp-ph-dropzone mb-2",style:o},r.createElement("div",{className:"comp-ph-dropzone-placeholder",dangerouslySetInnerHTML:{__html:a.dropzoneHtml}}),r.createElement("div",{id:`dragAnswer_${e.id}`,className:"comp-ph-draggable",style:l,dangerouslySetInnerHTML:{__html:e.text}})))))))})),bt=(0,We.PA)((e=>{const{question:t,getAnswers:n,onChanged:s}=e;if("combobox"!==t.options.displayMode)return null;const{groups:i=[]}=t,a=i.reduce(((e,t)=>t.text.length>e?t.text.length:e),0);return r.createElement("div",null,r.createElement("p",{className:"mb-5 comp-ph-question-text",dangerouslySetInnerHTML:{__html:t.text}}),r.createElement("div",null,t.answers.map((e=>{var t,o,l;return r.createElement("div",{className:"row mb-3"},r.createElement("div",{className:"col-md-6",dangerouslySetInnerHTML:{__html:e.text}}),r.createElement("div",{className:"col-md-auto"},r.createElement("div",{style:{width:8*a+100+"px"}},r.createElement(pt.Ay,{defaultValue:null!==(l=null===(o=null===(t=n().find((t=>t.answer[0]===e.id)))||void 0===t?void 0:t.answer)||void 0===o?void 0:o[1])&&void 0!==l?l:null,options:i.map((e=>({value:e.id,label:e.text}))),components:{Option:wt,SingleValue:yt},onChange:t=>{if(!t)return;const r=n().filter((t=>t.answer[0]!==e.id)),i=[...r,{answer:[e.id,+t.value],isСreatedByUser:!0}];s(i)}}))))}))))})),Et=(0,We.PA)((e=>{const{question:t,getAnswers:n,onChanged:s}=e;if("combobox"!==t.options.displayMode)return null;const{groups:i=[],options:a}=t;return(0,r.useEffect)((()=>{document.querySelectorAll(`#question_${t.questionId} [data-answer-id]`).forEach((e=>{const t=+e.getAttribute("data-answer-id"),a=r.createElement(pt.Ay,{options:i.map((e=>({value:e.id,label:e.text}))),components:{Option:wt,SingleValue:yt},onChange:e=>{if(!e)return;const r=n().filter((e=>e.answer[0]!==t)),i=[...r,{answer:[t,+e.value],isСreatedByUser:!0}];s(i)}});mt.render(a,e)}))}),[t.questionId]),r.createElement("div",{id:`question_${t.questionId}`},r.createElement("p",{className:"comp-ph-question-text",dangerouslySetInnerHTML:{__html:t.text}}))})),wt=e=>{const{innerRef:t,innerProps:n,children:s}=e,i=ht(e,["innerRef","innerProps","children"]);return r.createElement(gt.c.Option,Object.assign({},i),r.createElement("div",Object.assign({ref:t},n,{dangerouslySetInnerHTML:{__html:e.label}})))},yt=e=>{var t;const{innerRef:n,innerProps:s,children:i}=e,a=ht(e,["innerRef","innerProps","children"]);return r.createElement(gt.c.SingleValue,Object.assign({},a),null===(t=e.getValue())||void 0===t?void 0:t.map((e=>r.createElement("div",Object.assign({ref:n},s,{dangerouslySetInnerHTML:{__html:e.label}})))))};var xt=function(e,t){var n={};for(var s in e)Object.prototype.hasOwnProperty.call(e,s)&&t.indexOf(s)<0&&(n[s]=e[s]);if(null!=e&&"function"==typeof Object.getOwnPropertySymbols){var r=0;for(s=Object.getOwnPropertySymbols(e);r<s.length;r++)t.indexOf(s[r])<0&&Object.prototype.propertyIsEnumerable.call(e,s[r])&&(n[s[r]]=e[s[r]])}return n};const Ct=e=>{var{id:t,title:n,value:s,onChange:i,isChecked:a,style:o}=e,l=xt(e,["id","title","value","onChange","isChecked","style"]);return r.createElement("label",Object.assign({htmlFor:t,onClick:()=>i(s),style:a&&o||void 0},l),n)},St=e=>{var{id:t,value:n,name:s,selected:i}=e,a=xt(e,["id","value","name","selected"]);return r.createElement("input",Object.assign({id:t,type:"radio",name:s,checked:i===n,readOnly:!0},a))},_t=(0,We.PA)((e=>{const t=t=>{var n;null===(n=e.onChange)||void 0===n||n.call(e,t)};return r.createElement("div",{id:e.id,className:"switch-field d-inline-flex flex-row"},e.values.map(((n,s)=>{var i,a,o,l;return r.createElement("span",{key:s},r.createElement(St,Object.assign({id:`${e.id}_${n}_checkbox`,name:`${e.id}_switch`,"data-value":n,value:n,selected:e.selected},e.inputAttributes)),r.createElement(Ct,Object.assign({id:`${e.id}_${n}_checkbox`,isChecked:n===e.selected,title:null!==(a=null===(i=e.displayNames)||void 0===i?void 0:i[s])&&void 0!==a?a:n,value:n,onChange:t,style:null!==(l=null===(o=e.valueStyles)||void 0===o?void 0:o[s])&&void 0!==l?l:void 0},e.inputAttributes)))})))})),Nt=(0,We.PA)((e=>{const{question:t}=e,{options:n}=t;switch(!0){case"switch"==n.displayMode&&!n.requireContext:return r.createElement(At,Object.assign({},e));case"switch"===n.displayMode&&n.requireContext:return r.createElement(Ot,Object.assign({},e));case"dragNdrop"===n.displayMode:return r.createElement(It,Object.assign({},e))}return r.createElement("div",null,"Not implemented")})),At=(0,We.PA)((e=>{var t;const{question:n,getAnswers:s,onChanged:i}=e;if("switch"!==n.options.displayMode)return null;const{options:a}=n,o=null!==(t=a.selectorReplacers)&&void 0!==t?t:["no","yes"];return r.createElement("div",null,r.createElement("p",null,r.createElement("div",{className:"comp-ph-question-text",dangerouslySetInnerHTML:{__html:n.text}})),r.createElement("p",{className:"d-flex flex-column"},n.answers.map((e=>{var t,a,l,c;return r.createElement("div",{className:"d-flex flex-row mb-3"},r.createElement("div",{className:"mr-2 mt-1"},r.createElement(_t,{id:`question_${n.questionId}_anwser_${e.id}`,selected:null!==(c=o[null===(l=null===(a=null===(t=s().filter((t=>t.answer[0]===e.id)))||void 0===t?void 0:t[0])||void 0===a?void 0:a.answer)||void 0===l?void 0:l[1]])&&void 0!==c?c:"",values:o,inputAttributes:{"data-answer-id":e.id},onChange:t=>((e,t)=>{const n=o.indexOf(t),r=[...s().filter((t=>t.answer[0]!==e)),{answer:[e,n],isСreatedByUser:!0}];i(r)})(e.id,t)})),r.createElement("div",null,e.text))}))))})),Ot=(0,We.PA)((e=>{var t;const{question:n,getAnswers:s,onChanged:i}=e;if("switch"!==n.options.displayMode)return null;const{options:a}=n,o=null!==(t=a.selectorReplacers)&&void 0!==t?t:["no","yes"];return(0,r.useEffect)((()=>{document.querySelectorAll(`#question_${n.questionId} [data-answer-id]`).forEach((e=>{var t,n,a,l,c;const u=+(null!==(t=e.getAttribute("data-answer-id"))&&void 0!==t?t:-1),d=r.createElement(_t,{id:`toggle_answer_${u}`,selected:null!==(c=o[null===(l=null===(a=null===(n=s().filter((e=>e.answer[0]===+u)))||void 0===n?void 0:n[0])||void 0===a?void 0:a.answer)||void 0===l?void 0:l[1]])&&void 0!==c?c:"",inputAttributes:{"data-answer-id":u},values:o,onChange:e=>((e,t)=>{const n=o.indexOf(t),r=[...s().filter((t=>t.answer[0]!==e)),{answer:[e,n],isСreatedByUser:!0}];i(r)})(+u,e)});mt.render(d,e),e.id=""}))}),[n.questionId]),(0,r.useEffect)((()=>{document.querySelectorAll(`#question_${n.questionId} input[data-answer-id]`).forEach((e=>{e.getAttribute("data-answer-id"),e.checked=void 0})),s().forEach((({answer:e})=>{const[t,n]=e,s=document.querySelector(`#toggle_answer_${t} input[data-value='${o[n]}']`);s&&!s.checked&&setTimeout((()=>s.checked=!0),10)}))}),[n.questionId,s()]),r.createElement("div",{id:`question_${n.questionId}`},r.createElement("p",null,r.createElement("div",{className:"comp-ph-question-text",dangerouslySetInnerHTML:{__html:n.text}})))})),It=(0,We.PA)((e=>{const{question:t,getAnswers:n,onChanged:s,answers:i}=e;if("dragNdrop"!==t.options.displayMode)return null;const a=Object.assign(Object.assign({},t),{type:"MATCHING",options:Object.assign(Object.assign({},t.options),{multipleSelectionEnabled:!0}),groups:[{id:0,text:'<span class="badge badge-success" style="height: 100%; width: 100%;display: flex; align-items: center;justify-content: center;">✓</span>'},{id:1,text:'<span class="badge badge-danger" style="height: 100%; width: 100%;display: flex; align-items: center;justify-content: center;">x</span>'}]});return r.createElement(ft,{question:a,getAnswers:n,onChanged:s,answers:i})})),qt=(0,We.PA)((e=>{var t,n,s;const{question:i,getAnswers:a,onChanged:o,getFeedback:l}=e;if(!i.options.requireContext)return null;const{options:u}=i,d=null!==(t=u.orderNumberOptions)&&void 0!==t?t:{delimiter:"/",position:"SUFFIX"},m=(0,r.useMemo)((()=>{const e=document.createElement("div");return e.innerHTML=i.text,e}),[]);(0,r.useEffect)((()=>{document.querySelectorAll(`#question_${i.questionId} [data-answer-id]`).forEach((e=>{var t;const n=+(null!==(t=e.getAttribute("data-answer-id"))&&void 0!==t?t:"");e.addEventListener("click",(()=>o([...a(),{answer:[n,n],isСreatedByUser:!0}])))})),document.querySelectorAll("[data-comp-ph-value]").forEach((e=>{const t=e.getAttribute("data-comp-ph-value");e.innerHTML+=`<span class="comp-ph-expr-bottom-hint">${t}</span>`})),document.querySelectorAll("[data-comp-ph-pos]").forEach((e=>{const t=e.getAttribute("data-comp-ph-pos");e.innerHTML+=`<span class="comp-ph-expr-top-hint">${t}</span>`}))}),[i.questionId]),(0,r.useEffect)((()=>{var e,t;document.querySelectorAll(`#question_${i.questionId} [data-answer-id]`).forEach((e=>{var t;const n=e.getAttribute("data-comp-ph-value"),s=e.getAttribute("data-comp-ph-pos");e.innerHTML=(null===(t=m.querySelector(`#${e.id}`))||void 0===t?void 0:t.innerHTML)+(s?`<span class="comp-ph-expr-top-hint">${s}</span>`:"")+(n?`<span class="comp-ph-expr-bottom-hint">${n}</span>`:""),e.classList.remove("disabled"),e.classList.remove("comp-ph-question-answer--last-selected-by-system")}));const n=null===(t=null===(e=l())||void 0===e?void 0:e.isCorrect)||void 0===t||t?[...a()].reverse().reduce(((e,t)=>{const[n,s]=e;return-1===n?e:t.isСreatedByUser||0!==n&&n!==t.createdByInteraction?[-1,s]:(s.push(t),[t.createdByInteraction||-1,s])}),[0,[]])[1].reverse():[];a().forEach(((e,t,s)=>{const{answer:r}=e,[i]=r,a=document.querySelectorAll(`[data-answer-id='${i}']`);if(!a.length)return 0;a.forEach((s=>{var r,i;if(n.includes(e)&&s.classList.add("comp-ph-question-answer--last-selected-by-system"),"NONE"!==d.position){const e=d.delimiter,n=null!==(i=null===(r=d.replacers)||void 0===r?void 0:r[t])&&void 0!==i?i:t+1,a="PREFIX"===d.position?`${n}${e}${s.innerHTML}`:"SUFFIX"===d.position?`${s.innerHTML}${e}${n}`:"BOTTOM"===d.position?`<span class="comp-ph-expr-bottom-hint">${e}${n}</span>${s.innerHTML}`:s.innerHTML;s.innerHTML=a}u.multipleSelectionEnabled||s.classList.add("disabled")}))}))}),[i.questionId,a().length]);const p=null!==(s=null===(n=l())||void 0===n?void 0:n.trace)&&void 0!==s?s:0===a().length?i.initialTrace:null,g=u.showTrace&&!c(p)&&p.length>0;return r.createElement("div",{id:`question_${i.questionId}`},r.createElement("div",{className:"comp-ph-question-text",dangerouslySetInnerHTML:{__html:i.text}}),r.createElement(Ke,{isVisible:g},r.createElement("p",null,r.createElement("table",{className:"comp-ph-trace"},r.createElement("tbody",null,null==p?void 0:p.map(((e,t)=>r.createElement("tr",{key:t},r.createElement("td",{dangerouslySetInnerHTML:{__html:e}})))))))))})),kt=(0,We.PA)((e=>{const{question:t}=e;switch(!0){case"radio"===t.options.displayMode&&!t.options.requireContext:return r.createElement(Lt,Object.assign({},e));case"radio"===t.options.displayMode&&t.options.requireContext:return r.createElement(jt,Object.assign({},e))}return r.createElement("div",null,"Not implemented")})),Lt=(0,We.PA)((e=>{const{question:t,getAnswers:n,onChanged:s}=e;return"radio"!==t.options.displayMode?null:r.createElement("div",null,r.createElement("div",{className:"mb-3"},r.createElement("div",{className:"comp-ph-question-text",dangerouslySetInnerHTML:{__html:t.text}})),r.createElement("div",{className:"d-flex flex-column"},t.answers.map(((e,i)=>r.createElement("label",{htmlFor:`question_${t.questionId}_answer_${e.id}`,className:"comp-ph-singlechoice-label d-flex flex-row "+(i!==t.answers.length-1?"mb-3":"")},r.createElement("div",{className:"mr-2 mt-1"},r.createElement("input",{id:`question_${t.questionId}_answer_${e.id}`,name:`switch_${t.questionId}`,type:"radio",checked:n().some((t=>t.answer[0]===e.id)),onChange:t=>{return n=e.id,void(t.target.checked&&s([{answer:[n,n],isСreatedByUser:!0}]));var n},readOnly:!0})),r.createElement("div",null,e.text))))))})),jt=(0,We.PA)((e=>{const{question:t,getAnswers:n,onChanged:s}=e;if("radio"!==t.options.displayMode)return null;const{options:i}=t;return(0,r.useEffect)((()=>{document.querySelectorAll(`#question_${t.questionId} [data-answer-id]`).forEach((e=>{var i;const a=null!==(i=e.getAttribute("data-answer-id"))&&void 0!==i?i:-1,o=r.createElement("label",{htmlFor:`question_${t.questionId}_answer_${a}`,className:"comp-ph-singlechoice-label"},r.createElement("input",{id:`question_${t.questionId}_answer_${a}`,name:`switch_${t.questionId}`,"data-answer-id":a,type:"radio",checked:n().some((e=>e.answer[0]===+a)),onChange:e=>{return t=+a,void(e.target.checked&&s([{answer:[t,t],isСreatedByUser:!0}]));var t},readOnly:!0}),r.createElement("span",{dangerouslySetInnerHTML:{__html:e.innerHTML}}));mt.render(o,e),e.id=""}))}),[t.questionId]),(0,r.useEffect)((()=>{document.querySelectorAll(`#question_${t.questionId} input[data-answer-id]`).forEach((e=>{e.checked=void 0})),n().forEach((({answer:e})=>{const n=e[0],s=document.querySelector(`#question_${t.questionId} input[data-answer-id='${n}']`);s&&setTimeout((()=>s.checked=!0),10)}))}),[t.questionId,n()]),r.createElement("div",{id:`question_${t.questionId}`},r.createElement("p",null,r.createElement("div",{className:"comp-ph-question-text",dangerouslySetInnerHTML:{__html:t.text}})))})),Qt=(0,We.PA)((e=>{var t;const{question:n,answers:s,onChanged:i,getAnswers:a,getFeedback:o,isFeedbackLoading:l,isQuestionFreezed:c}=e;let u;switch(n.type){case"MATCHING":u=r.createElement(vt,{question:n,onChanged:i,answers:s,getAnswers:a});break;case"MULTI_CHOICE":u=r.createElement(Nt,{question:n,onChanged:i,answers:s,getAnswers:a});break;case"SINGLE_CHOICE":u=r.createElement(kt,{question:n,onChanged:i,answers:s,getAnswers:a});break;case"ORDER":u=r.createElement(qt,{question:n,onChanged:i,answers:s,getAnswers:a,getFeedback:o});break;default:return(0,I.absurd)(n)}const d=["comp-ph-question-wrapper",0===(null===(t=o())||void 0===t?void 0:t.stepsLeft)?"comp-ph-question-wrapper--finished":"",l?"comp-ph-question-wrapper--loading-feedback":"",c?"comp-ph-question-wrapper--freezed":""].join(" ");return r.createElement("div",{className:d},u)}));var Tt=n(5677),Pt=function(e,t,n,s){return new(n||(n=Promise))((function(r,i){function a(e){try{l(s.next(e))}catch(e){i(e)}}function o(e){try{l(s.throw(e))}catch(e){i(e)}}function l(e){var t;e.done?r(e.value):(t=e.value,t instanceof n?t:new n((function(e){e(t)}))).then(a,o)}l((s=s.apply(e,t||[])).next())}))};const Dt=(0,We.PA)((e=>{const{violationLaw:t,store:n}=e,[s,i]=(0,r.useState)(!1),[a,o]=(0,r.useState)(!0),[c,u]=(0,r.useState)(!0),[d,m]=(0,r.useState)(t),{t:p}=(0,qe.Bd)(),g=()=>Pt(void 0,void 0,void 0,(function*(){var e;"COMPLETED"!=n.questionState&&n.question||"CONTINUE_AUTO"!==(null===(e=n.feedback)||void 0===e?void 0:e.action)||(console.log("show feedback for 3 seconds"),yield l(3e3),yield h())})),h=()=>Pt(void 0,void 0,void 0,(function*(){var e,t;const s=(null===(t=null===(e=n.feedback)||void 0===e?void 0:e.message)||void 0===t?void 0:t.violationLaws)||null;if(!s)return console.log("empty violation laws"),void u(!1);m(s),yield n.generateSupplementaryQuestion(s.map((e=>e.name))),yield g()}));return r.createElement(Ke,{isVisible:c},r.createElement(Ke,{isVisible:a},r.createElement("div",{className:"d-flex flex-row mt-3"},r.createElement(Ue.A,{onClick:()=>Pt(void 0,void 0,void 0,(function*(){var e;o(!1),i(!0),yield n.generateSupplementaryQuestion(d.map((e=>e.name))),n.question&&"FINISH"!==(null===(e=n.feedback)||void 0===e?void 0:e.action)||(console.log("no need to generate sup question"),u(!1))})),variant:"primary"},p("exercise_supquestion_details")),r.createElement(Ue.A,{onClick:()=>{u(!1)},variant:"success",className:"ml-2"},p("exercise_supquestion_gotit")))),r.createElement(Ze,{type:"DIALOG",size:"xl",show:s,closeButton:!1,handleClose:()=>i(!1)},r.createElement(Rt,{store:n,onSubmitted:()=>Pt(void 0,void 0,void 0,(function*(){yield n.sendAnswers();const{feedback:e}=n;if(!e)return console.log("empty feedback for question asnwer"),void u(!1);yield g()})),onNextQuestionRequested:h})))})),Rt=(0,We.PA)((e=>{var t,n;const{store:s,onSubmitted:i,onNextQuestionRequested:a}=e,{t:o}=(0,qe.Bd)(),l=s.question;if("LOADING"===s.questionState)return r.createElement("div",{className:"mt-2"},r.createElement(lt,null));const c="EXPLICIT"===s.questionSubmitMode&&s.canSendQuestionAnswers,u="COMPLETED"===s.questionState&&!!s.feedback&&!!l,d="COMPLETED"===s.questionState&&!!s.feedback&&!l,m="CONTINUE_MANUAL"===(null===(t=s.feedback)||void 0===t?void 0:t.action)&&(u||d)&&!!(null===(n=s.feedback)||void 0===n?void 0:n.message.violationLaws);return r.createElement(r.Fragment,null,l&&r.createElement(Qt,{question:l,answers:s.answer,getAnswers:()=>s.answer,onChanged:e=>{s.setAnswer(e),"IMPLICIT"===s.questionSubmitMode&&(null==i||i())},getFeedback:()=>{},isFeedbackLoading:s.isFeedbackLoading,isQuestionFreezed:s.isQuestionFreezed}),s.isFeedbackLoading&&r.createElement("div",{className:"mt-2"},r.createElement(lt,null)),c&&r.createElement(Ue.A,{variant:"primary",onClick:i},o("exercise_supquestion_send_answer")),d&&r.createElement(r.Fragment,null,s.feedback.message.message),u&&r.createElement("div",{className:"mt-2"},r.createElement(Mt,{message:s.feedback.message})),m&&r.createElement("div",{className:"mt-3"},r.createElement(Ue.A,{variant:"primary",onClick:a},o("exercise_supquestion_next_question"))))})),Mt=(0,We.PA)((e=>{let{supQuestionStore:t,message:n,showGenerateSupQuestion:s}=e;s=s&&null!=t;const i="SUCCESS"===n.type?"success":"danger";return r.createElement(Tt.A,{variant:i},n.message)}));var $t=n(6052);const Ht=(0,We.PA)((({store:e,showExtendedFeedback:t})=>{const{feedback:n,isFeedbackVisible:s,question:i}=e,a="ANSWER_EVALUATING"===e.questionState,o="LOADING"===e.questionState,{t:l}=(0,qe.Bd)();if(a)return r.createElement("div",{className:"mt-2"},r.createElement(lt,null));if(!n||o||!i)return null;const u="COMPLETED"===e.questionState?[{type:"SUCCESS",message:l("issolved_feeback")}]:n.messages;return r.createElement("div",{className:"comp-ph-feedback-wrapper mt-2"},s&&r.createElement(r.Fragment,null,r.createElement("div",{className:"mb-3"},null==u?void 0:u.map((n=>{var s;return r.createElement(Ft,{message:n,supQuestionStore:e.supplementaryQuestion,showGenerateSupQuestion:t&&i.options.showSupplementaryQuestions&&"ERROR"===n.type&&(null===(s=n.violationLaws)||void 0===s?void 0:s.every((e=>e.canCreateSupplementaryQuestion)))})}))),t&&r.createElement("div",null,null!==n.grade&&r.createElement(r.Fragment,null,r.createElement($t.A,{variant:"primary"},l("grade_feeback"),": ",n.grade)," "),null!==n.correctSteps&&r.createElement(r.Fragment,null,r.createElement($t.A,{variant:"success"},l("correctsteps_feeback"),": ",n.correctSteps)," "),!c(n.stepsWithErrors)&&n.stepsWithErrors>0&&r.createElement(r.Fragment,null,r.createElement($t.A,{variant:"danger"},l("stepswitherrors_feeback"),": ",n.stepsWithErrors)," "),!c(n.stepsLeft)&&n.stepsLeft>0&&r.createElement(r.Fragment,null,r.createElement($t.A,{variant:"info"},l("stepsleft_feeback"),": ",n.stepsLeft)," "))))})),Ft=(0,We.PA)((e=>{var t;let{supQuestionStore:n,message:s,showGenerateSupQuestion:i}=e;i=i&&null!=n;const a="SUCCESS"===s.type?"success":"danger";return r.createElement(Tt.A,{variant:a},r.createElement("div",{"data-domain-laws":null===(t=s.violationLaws)||void 0===t?void 0:t.map((e=>e.name)).join(";"),dangerouslySetInnerHTML:{__html:s.message}}),i&&"ERROR"===s.type&&s.violationLaws&&r.createElement(Dt,{store:n,violationLaw:s.violationLaws})||null)}));const Vt=(0,We.PA)((e=>{const{store:t,showExtendedFeedback:n,onChanged:s}=e,i=t.question;return"LOADING"===t.questionState?r.createElement("div",{className:"mt-2"},r.createElement(lt,null)):i?r.createElement(r.Fragment,null,r.createElement(Qt,{question:i,answers:t.lastAnswer,getAnswers:()=>t.lastAnswer,onChanged:e=>{return n=void 0,r=void 0,a=function*(){(yield t.setFullAnswer(e))&&(null==s||s(e))},new((i=void 0)||(i=Promise))((function(e,t){function s(e){try{l(a.next(e))}catch(e){t(e)}}function o(e){try{l(a.throw(e))}catch(e){t(e)}}function l(t){var n;t.done?e(t.value):(n=t.value,n instanceof i?n:new i((function(e){e(n)}))).then(s,o)}l((a=a.apply(n,r||[])).next())}));var n,r,i,a},getFeedback:()=>t.feedback,isFeedbackLoading:"ANSWER_EVALUATING"===t.questionState,isQuestionFreezed:t.isQuestionFreezed}),r.createElement(Ht,{store:t,showExtendedFeedback:n})):null})),Gt=(0,We.PA)((()=>{var e,t;const[n]=(0,r.useState)((()=>i.kL.resolve(Ie)));return r.createElement(Vt,{store:n.currentQuestion,showExtendedFeedback:null===(t=null===(e=n.exercise)||void 0===e?void 0:e.options.supplementaryQuestionsEnabled)||void 0===t||t})}));var Bt=n(3133);const Yt=e=>{const{survey:t,enabledSurveyQuestions:n,isCompleted:s}=e,[a]=(0,r.useState)((()=>i.kL.resolve(fe))),[o,l]=(0,r.useState)(s?"COMPLETED":"INITAL"),[c,u]=(0,r.useState)(e.value||{}),{t:d}=(0,qe.Bd)(),m=e.survey.questions.filter((e=>n.includes(e.id)));var p=(e,t)=>{const n=Object.assign(Object.assign({},c),{[e]:t});u(n),console.log(n)};return m&&m.length?r.createElement("div",{className:"alert alert-warning rounded",role:"alert"},m.map(((e,t,n)=>r.createElement("div",{key:t,className:t!==n.length-1?"mb-4":""},"yes-no"===e.type&&r.createElement(zt,{isCompleted:s,question:e,onAnswered:p,value:c[e.id]})||"single-choice"===e.type&&r.createElement(Wt,{isCompleted:s,question:e,onAnswered:p,value:c[e.id]})||"open-ended"===e.type&&r.createElement(Ut,{isCompleted:s,question:e,onAnswered:p,value:c[e.id]})||"invalid question type"))),"SENDING_RESULTS"===o&&r.createElement("div",{className:"mt-2"},r.createElement(lt,{delay:100})),"VALIDATION_ERROR"===o&&r.createElement("div",{className:"mt-2"},r.createElement("div",{className:"alert alert-danger rounded"},"Необходимо ответить на все обязательные вопросы")),"COMPLETED"!==o&&"SENDING_RESULTS"!==o&&r.createElement("div",{className:"mt-2"},r.createElement(Ue.A,{variant:"primary",onClick:()=>{var n,s,r,i;n=void 0,s=void 0,i=function*(){m.filter((e=>e.required)).map((e=>e.id)).every((e=>c[e]))?(l("SENDING_RESULTS"),yield Promise.all(m.map((t=>a.postSurveyAnswer(t.id,e.questionId,c[t.id])))),l("COMPLETED"),e.onAnswersSended(t,e.questionId,c)):l("VALIDATION_ERROR")},new((r=void 0)||(r=Promise))((function(e,t){function a(e){try{l(i.next(e))}catch(e){t(e)}}function o(e){try{l(i.throw(e))}catch(e){t(e)}}function l(t){var n;t.done?e(t.value):(n=t.value,n instanceof r?n:new r((function(e){e(n)}))).then(a,o)}l((i=i.apply(n,s||[])).next())}))}},d("survey_sendresults")))):null},zt=e=>{const{question:t,onAnswered:n,value:s,isCompleted:i}=e;return r.createElement("div",null,r.createElement("div",{className:"mb-1"},t.text),r.createElement("div",{className:"d-flex flex-row mt-2"},r.createElement("button",{type:"button",className:"btn btn-secondary mr-2 "+(t.options.yesValue===s?"active":""),onClick:()=>n(t.id,t.options.yesValue),disabled:void 0!==s},t.options.yesText),r.createElement("button",{type:"button",className:"btn btn-secondary "+(t.options.noValue===s?"active":""),onClick:()=>n(t.id,t.options.noValue),disabled:void 0!==s},t.options.noText)))},Wt=e=>{const{question:t,onAnswered:n,value:s,isCompleted:i}=e;return r.createElement("div",null,r.createElement("div",{className:"mb-1"},t.text),r.createElement("div",{className:"d-flex mt-2"},r.createElement("div",null,t.options.map(((e,a,o)=>r.createElement("div",{key:e.id},r.createElement(Bt.A.Check,{disabled:i,checked:s===e.id,name:`radio_qid${t.id}`,type:"radio",id:`radio_qid${t.id}_sqid${e.id}`,label:e.text,value:e.id,onChange:e=>n(t.id,e.target.value)})))))))},Ut=e=>{const{question:t,onAnswered:n,value:s,isCompleted:i}=e;return r.createElement("div",null,r.createElement("div",{className:"mb-1"},t.text),r.createElement("div",{className:"d-flex flex-row mt-2"},r.createElement(Bt.A.Control,{disabled:i,value:s,as:"textarea",rows:3,onChange:e=>n(t.id,e.target.value)})))};var Xt=function(e,t,n,s){return new(n||(n=Promise))((function(r,i){function a(e){try{l(s.next(e))}catch(e){i(e)}}function o(e){try{l(s.throw(e))}catch(e){i(e)}}function l(e){var t;e.done?r(e.value):(t=e.value,t instanceof n?t:new n((function(e){e(t)}))).then(a,o)}l((s=s.apply(e,t||[])).next())}))};const Kt=(0,We.PA)((()=>{var e,t,n,s,a,o,l,c,u,d,m,p,g,h;const[v]=(0,r.useState)((()=>i.kL.resolve(Ie))),{exerciseState:f,setExerciseState:b,storeState:E,currentQuestion:w,survey:y}=v,{storeState:x}=w,{t:C}=(0,qe.Bd)();(0,r.useEffect)((()=>{Xt(void 0,void 0,void 0,(function*(){var e;if("LAUNCH_ERROR"===f)return;if(v.currentQuestion.question)return void b("EXERCISE");if(yield v.loadSessionInfo(),v.isDebug)return b("EXERCISE"),void N();const t=v.currentAttemptId;return t&&Number.isInteger(t)?(b("EXERCISE"),void A(+t)):(null===(e=v.exercise)||void 0===e?void 0:e.options.forceNewAttemptCreationEnabled)||!(yield v.loadExistingExerciseAttempt())?(b("EXERCISE"),void _()):void b("MODAL")}))}),[]);const S=(0,r.useCallback)((()=>{Xt(void 0,void 0,void 0,(function*(){var e,t,n;if(null===(e=v.currentAttempt)||void 0===e?void 0:e.questionIds.length){const e=null===(t=v.currentAttempt)||void 0===t?void 0:t.questionIds.length;yield v.currentQuestion.loadQuestion(null===(n=v.currentAttempt)||void 0===n?void 0:n.questionIds[e-1])}else yield v.generateQuestion()}))}),[v]),_=(0,r.useCallback)((()=>{Xt(void 0,void 0,void 0,(function*(){v.currentQuestion.setQuestionState("LOADING"),yield v.createExerciseAttempt(),S()}))}),[v]),N=(0,r.useCallback)((()=>{Xt(void 0,void 0,void 0,(function*(){v.currentQuestion.setQuestionState("LOADING"),yield v.createDebugExerciseAttempt(),S()}))}),[v]),A=(0,r.useCallback)((e=>{Xt(void 0,void 0,void 0,(function*(){v.currentQuestion.setQuestionState("LOADING"),yield v.loadExerciseAttempt(e),S()}))}),[v]),O=(0,r.useCallback)(((e,t,n)=>{v.setSurveyAnswers(t,n)}),[v]);return r.createElement(r.Fragment,null,r.createElement("div",{className:`compph-exercise ${v.isDebug&&"compph-exercise--debug"}`||""},r.createElement(ct,{isLoading:!0===v.isSessionLoading||"INITIAL"===f},r.createElement(Ke,{isVisible:"EXERCISE"===f||"COMPLETED"===f},r.createElement(at,null),r.createElement("div",{className:"mt-5"},r.createElement(Gt,null),null!=y&&("COMPLETED"===v.currentQuestion.questionState||"COMPLETED"===f)&&r.createElement("div",{className:"mt-2"},r.createElement(Yt,{questionId:null!==(t=null===(e=v.currentQuestion.question)||void 0===e?void 0:e.questionId)&&void 0!==t?t:-1,survey:y.survey,enabledSurveyQuestions:v.ensureQuestionSurveyExists(null!==(s=null===(n=w.question)||void 0===n?void 0:n.questionId)&&void 0!==s?s:-1),value:null===(l=y.questions[null!==(o=null===(a=v.currentQuestion.question)||void 0===a?void 0:a.questionId)&&void 0!==o?o:-1])||void 0===l?void 0:l.results,onAnswersSended:O,isCompleted:"COMPLETED"===(null===(d=y.questions[null!==(u=null===(c=v.currentQuestion.question)||void 0===c?void 0:c.questionId)&&void 0!==u?u:-1])||void 0===d?void 0:d.status)})),r.createElement(Ke,{isVisible:"EXERCISE"===f},r.createElement(Ke,{isVisible:"LOADED"===v.currentQuestion.questionState},r.createElement("div",{className:"mt-3"},r.createElement(ut,null))),r.createElement(Ke,{isVisible:null==y&&((null===(m=v.exercise)||void 0===m?void 0:m.options.newQuestionGenerationEnabled)||"COMPLETED"===v.currentQuestion.questionState)||null!=y&&"COMPLETED"===(null===(h=y.questions[null!==(g=null===(p=v.currentQuestion.question)||void 0===p?void 0:p.questionId)&&void 0!==g?g:-1])||void 0===h?void 0:h.status)},r.createElement("div",{className:"mt-2"},r.createElement(tt,null)))),r.createElement(Ke,{isVisible:"COMPLETED"===f},r.createElement("div",{className:"mt-3"},r.createElement(Tt.A,{variant:"success"},C("exercise_completed")))))),r.createElement(Ke,{isVisible:"MODAL"===f},r.createElement(Ze,{type:"DIALOG",title:C("foundExisitingAttempt_title"),primaryBtnTitle:C("foundExisitingAttempt_continueattempt"),handlePrimaryBtnClicked:()=>{b("EXERCISE"),S()},secondaryBtnTitle:C("foundExisitingAttempt_newattempt"),handleSecondaryBtnClicked:()=>{b("EXERCISE"),_()}},r.createElement("p",null,C("foundExisitingAttempt_descr"),"?")))),[E,x].filter((e=>"ERROR"===e.tag)).map(((e,t,n)=>"ERROR"===e.tag&&r.createElement("div",{className:"mt-2"},r.createElement(Tt.A,{variant:"danger"},e.error.message))))))}));const Zt=()=>{const[e,t]=(0,r.useState)([]),[n,s]=(0,r.useState)(!1);return(0,r.useEffect)((()=>{var e,n,r,a;e=void 0,n=void 0,a=function*(){const e=new URLSearchParams(window.location.search).get("exerciseId");if(null===e||Number.isNaN(+e))throw new Error("Invalid exerciseId url param");const n=i.kL.resolve(S);s(!0);const r=yield n.getExerciseStatistics(+e);g.isRight(r)&&t(r.right),s(!1)},new((r=void 0)||(r=Promise))((function(t,s){function i(e){try{l(a.next(e))}catch(e){s(e)}}function o(e){try{l(a.throw(e))}catch(e){s(e)}}function l(e){var n;e.done?t(e.value):(n=e.value,n instanceof r?n:new r((function(e){e(n)}))).then(i,o)}l((a=a.apply(e,n||[])).next())}))}),[]),r.createElement("div",null,r.createElement("table",{className:"table"},r.createElement("thead",null,r.createElement("tr",null,r.createElement("th",{scope:"col"},"AttemptId"),r.createElement("th",{scope:"col"},"QuestionsCount"),r.createElement("th",{scope:"col"},"TotalInteractionsCount"),r.createElement("th",{scope:"col"},"TotalInteractionsWithErrorsCount"),r.createElement("th",{scope:"col"},"AverageGrade"))),r.createElement("tbody",null,r.createElement(ct,{isLoading:n},e.map(((e,t)=>r.createElement("tr",{key:t},r.createElement("th",{scope:"row"},e.attemptId),r.createElement("td",null,e.questionsCount),r.createElement("td",null,e.totalInteractionsCount),r.createElement("td",null,e.totalInteractionsWithErrorsCount),r.createElement("td",null,e.averageGrade))))))))};var Jt=n(2374);const en=(0,We.PA)((()=>{const[e,t]=(0,r.useState)([]),[n,s]=(0,r.useState)(!1);return(0,r.useEffect)((()=>{var e,n,r,a;e=void 0,n=void 0,a=function*(){s(!0);const e=i.kL.resolve(S),n=yield e.getExercises();g.isRight(n)&&t(n.right),s(!1)},new((r=void 0)||(r=Promise))((function(t,s){function i(e){try{l(a.next(e))}catch(e){s(e)}}function o(e){try{l(a.throw(e))}catch(e){s(e)}}function l(e){var n;e.done?t(e.value):(n=e.value,n instanceof r?n:new r((function(e){e(n)}))).then(i,o)}l((a=a.apply(e,n||[])).next())}))}),[]),r.createElement("div",null,r.createElement(Jt.A,null,e.map((e=>r.createElement(Jt.A.Item,null,r.createElement("a",{href:`exercise?exerciseId=${e}`},e))))))}));var tn=function(e,t,n,s){return new(n||(n=Promise))((function(r,i){function a(e){try{l(s.next(e))}catch(e){i(e)}}function o(e){try{l(s.throw(e))}catch(e){i(e)}}function l(e){var t;e.done?r(e.value):(t=e.value,t instanceof n?t:new n((function(e){e(t)}))).then(a,o)}l((s=s.apply(e,t||[])).next())}))};const nn=(0,We.PA)((()=>{var e,t,n,s,a,o,l,c,u,d,m;const[p]=(0,r.useState)((()=>i.kL.resolve(Ie))),{exerciseState:g,setExerciseState:h,storeState:v,currentQuestion:f,survey:b}=p,{storeState:E}=f,{t:w}=(0,qe.Bd)(),[y,x]=(0,r.useState)("ACTIVE");(0,r.useEffect)((()=>{tn(void 0,void 0,void 0,(function*(){"LAUNCH_ERROR"!==g&&(p.currentQuestion.question?h("EXERCISE"):(yield p.loadSessionInfo(),h("EXERCISE"),yield C()))}))}),[]);const C=()=>tn(void 0,void 0,void 0,(function*(){p.currentQuestion.setQuestionState("LOADING"),yield p.createExerciseAttempt(),yield tn(void 0,void 0,void 0,(function*(){var e,t,n;if(null===(e=p.currentAttempt)||void 0===e?void 0:e.questionIds.length){const e=null===(t=p.currentAttempt)||void 0===t?void 0:t.questionIds.length;yield p.currentQuestion.loadQuestion(null===(n=p.currentAttempt)||void 0===n?void 0:n.questionIds[e-1])}else yield p.generateQuestion()}))})),S=(0,r.useCallback)(((e,t,n)=>{console.log("лень рефакторить, не работает крч")}),[p]),_=null===(e=p.exercise)||void 0===e?void 0:e.options.surveyOptions;return null===(t=p.currentQuestion.question)||void 0===t||t.questionId,(null==_?void 0:_.enabled)?r.createElement(r.Fragment,null,r.createElement(ct,{isLoading:!0===p.isSessionLoading||"INITIAL"===g},r.createElement(Ke,{isVisible:"EXERCISE"===g||"COMPLETED"===g},r.createElement(at,null),r.createElement("div",{className:"mt-5"},r.createElement("div",{style:{pointerEvents:"none"}},r.createElement(Gt,null)),r.createElement(Ke,{isVisible:"COMPLETED"===y},r.createElement("div",{className:"mt-2"},r.createElement(Tt.A,{variant:"success"},"Спасибо за участие в опросе!"))),r.createElement(Ke,{isVisible:"COMPLETED"!==y&&"LOADED"===f.questionState},r.createElement("div",{className:"mt-2"},r.createElement(Yt,{questionId:null!==(s=null===(n=p.currentQuestion.question)||void 0===n?void 0:n.questionId)&&void 0!==s?s:-1,survey:p.survey.survey,value:null===(a=p.survey)||void 0===a?void 0:a.questions[null!==(l=null===(o=p.currentQuestion.question)||void 0===o?void 0:o.questionId)&&void 0!==l?l:-1].results,enabledSurveyQuestions:null!==(m=null===(c=p.survey)||void 0===c?void 0:c.questions[null!==(d=null===(u=p.currentQuestion.question)||void 0===u?void 0:u.questionId)&&void 0!==d?d:-1].questions)&&void 0!==m?m:[],onAnswersSended:S})))))),[v,E].filter((e=>"ERROR"===e.tag)).map(((e,t,n)=>"ERROR"===e.tag&&r.createElement("div",{className:"mt-2"},r.createElement(Tt.A,{variant:"danger"},e.error.message))))):null}));var sn=function(e,t){return function(n,s){t(n,s,e)}},rn=function(e,t,n,s){return new(n||(n=Promise))((function(r,i){function a(e){try{l(s.next(e))}catch(e){i(e)}}function o(e){try{l(s.throw(e))}catch(e){i(e)}}function l(e){var t;e.done?r(e.value):(t=e.value,t instanceof n?t:new n((function(e){e(t)}))).then(a,o)}l((s=s.apply(e,t||[])).next())}))};class an{constructor(e,t,n){Object.defineProperty(this,"exerciseSettingsController",{enumerable:!0,configurable:!0,writable:!0,value:e}),Object.defineProperty(this,"card",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"concepts",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"laws",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"skills",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"numberOfQuestions",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"bankLoadingState",{enumerable:!0,configurable:!0,writable:!0,value:"NOT_STARTED"}),Object.defineProperty(this,"bankSearchResult",{enumerable:!0,configurable:!0,writable:!0,value:{questions:[],count:0,topRatedCount:0}}),Object.defineProperty(this,"complexity",{enumerable:!0,configurable:!0,writable:!0,value:.5}),Object.defineProperty(this,"autorunner",{enumerable:!0,configurable:!0,writable:!0,value:void 0}),Object.defineProperty(this,"abortController",{enumerable:!0,configurable:!0,writable:!0,value:null}),this.concepts=n.concepts,this.laws=n.laws,this.skills=n.skills,this.numberOfQuestions=n.numberOfQuestions,this.complexity=n.complexity,this.card=t,(0,O.l_)(this),this.autorunner=(0,O.fm)((()=>rn(this,void 0,void 0,(function*(){const e=this.complexity,n=this.laws.slice(),s=this.concepts.slice(),r=this.skills.slice();this.updateBankStats(s,n,r,t.tags,e)}))),{delay:1e3})}*updateBankStats(e,t,n,s,r){const{card:i}=this;this.abortController&&(this.abortController.abort(),this.abortController=null);const a=new AbortController;this.abortController=a,(0,O.h5)((()=>this.bankLoadingState="IN_PROGRESS"));const o=yield this.exerciseSettingsController.search(i.domainId,e,t,n,s,r,5,a.signal);g.isRight(o)&&((0,O.h5)((()=>{this.bankSearchResult=o.right})),(0,O.h5)((()=>this.bankLoadingState="COMPLETED"))),this.abortController===a&&(this.abortController=null)}[Symbol.dispose](){this.autorunner&&this.autorunner(),this.abortController&&(this.abortController.abort(),this.abortController=null)}}let on=class{constructor(e,t,n){Object.defineProperty(this,"exerciseSettingsController",{enumerable:!0,configurable:!0,writable:!0,value:e}),Object.defineProperty(this,"userController",{enumerable:!0,configurable:!0,writable:!0,value:t}),Object.defineProperty(this,"exerciseController",{enumerable:!0,configurable:!0,writable:!0,value:n}),Object.defineProperty(this,"exercisesLoadStatus",{enumerable:!0,configurable:!0,writable:!0,value:"NONE"}),Object.defineProperty(this,"exercises",{enumerable:!0,configurable:!0,writable:!0,value:null}),Object.defineProperty(this,"domains",{enumerable:!0,configurable:!0,writable:!0,value:null}),Object.defineProperty(this,"backends",{enumerable:!0,configurable:!0,writable:!0,value:null}),Object.defineProperty(this,"strategies",{enumerable:!0,configurable:!0,writable:!0,value:null}),Object.defineProperty(this,"currentCard",{enumerable:!0,configurable:!0,writable:!0,value:null}),Object.defineProperty(this,"user",{enumerable:!0,configurable:!0,writable:!0,value:null}),Object.defineProperty(this,"changeLanguage",{enumerable:!0,configurable:!0,writable:!0,value:e=>{this.user&&this.user.language!==e&&(this.user.language=e,le.Ay.changeLanguage(e))}}),(0,O.l_)(this)}toCardViewModel(e){var t;const n=null===(t=this.domains)||void 0===t?void 0:t.find((t=>t.id===e.domainId));if(!n)throw new Error(`не найден домен ${e.domainId}`);const s=(0,O.sH)(Object.assign(Object.assign({},e),{tags:e.tags.filter((e=>n.tags.some((t=>t===e)))),stages:[]}));return s.stages=(0,I.pipe)(e.stages,be.map((e=>new an(this.exerciseSettingsController,s,e)))),s}fromCardViewModel(e){return Object.assign(Object.assign({},e),{stages:(0,I.pipe)(e.stages,be.map((e=>({concepts:e.concepts,laws:e.laws,skills:e.skills,numberOfQuestions:e.numberOfQuestions,complexity:e.complexity}))))})}loadExercises(){return rn(this,void 0,void 0,(function*(){if("LOADED"===this.exercisesLoadStatus||"LOADING"===this.exercisesLoadStatus)return;(0,O.h5)((()=>this.exercisesLoadStatus="LOADING"));const[e,t,n,s,r]=yield Promise.all([this.exerciseSettingsController.getAllExercises(),this.exerciseSettingsController.getDomains(),this.exerciseSettingsController.getBackends(),this.exerciseSettingsController.getStrategies(),this.userController.getCurrentUser()]);g.isRight(e)&&g.isRight(t)&&g.isRight(n)&&g.isRight(s)&&g.isRight(r)&&(0,O.h5)((()=>{this.exercises=e.right,this.domains=t.right,this.backends=n.right,this.strategies=s.right,this.user=r.right,le.Ay.changeLanguage(r.right.language)})),(0,O.h5)((()=>this.exercisesLoadStatus="LOADED"))}))}loadExercise(e){return rn(this,void 0,void 0,(function*(){if("LOADED"!==this.exercisesLoadStatus)throw new Error("Exercises must be loaded first");(0,O.h5)((()=>this.exercisesLoadStatus="EXERCISELOADING"));const t=yield this.exerciseSettingsController.getExercise(e);g.isRight(t)&&(0,O.h5)((()=>{this.currentCard=this.toCardViewModel(t.right)})),(0,O.h5)((()=>this.exercisesLoadStatus="LOADED"))}))}createNewExecise(){return rn(this,void 0,void 0,(function*(){if("LOADED"!==this.exercisesLoadStatus)throw new Error("Exercises must be loaded first");const e=yield this.exerciseSettingsController.createExercise("(empty)",this.domains[0].id,this.strategies[0].id);if(!g.isRight(e))return;(0,O.h5)((()=>this.exercisesLoadStatus="EXERCISELOADING"));const[t,n]=yield Promise.all([this.exerciseSettingsController.getExercise(e.right),this.exerciseSettingsController.getAllExercises()]);g.isRight(t)&&g.isRight(n)&&(0,O.h5)((()=>{this.currentCard=this.toCardViewModel(t.right),this.exercises=n.right})),(0,O.h5)((()=>this.exercisesLoadStatus="LOADED"))}))}saveCard(){return rn(this,void 0,void 0,(function*(){if(!this.currentCard)return;(0,O.h5)((()=>this.exercisesLoadStatus="EXERCISELOADING")),yield this.exerciseSettingsController.saveExercise(this.fromCardViewModel(this.currentCard));const e=yield this.exerciseSettingsController.getAllExercises();g.isRight(e)&&(0,O.h5)((()=>{this.exercises=e.right})),(0,O.h5)((()=>this.exercisesLoadStatus="LOADED"))}))}setCardName(e){this.currentCard&&(this.currentCard.name=e)}setCardDomain(e){this.currentCard&&e!==this.currentCard.domainId&&(this.currentCard.stages[0].laws=[],this.currentCard.stages[0].concepts=[],this.currentCard.stages.splice(1),this.currentCard.domainId=e)}setCardStrategy(e){this.currentCard&&this.currentCard.strategyId!==e&&(this.currentCard.stages[0].laws=[],this.currentCard.stages[0].concepts=[],this.currentCard.stages.splice(1),this.currentCard.strategyId=e)}setCardStageComplexity(e,t){if(!this.currentCard||!this.currentCard.stages[e])return;const n=this.currentCard.stages[e],s=Number.parseInt(t);n.complexity=s/100}setCardCommonConceptValue(e,t){if(this.currentCard)for(const n of this.currentCard.stages){const s=n.concepts.findIndex((t=>t.name==e));let r=-1!==s?n.concepts[s]:null;"PERMITTED"!==t?r?n.concepts[s]=Object.assign(Object.assign({},r),{kind:t}):(r={name:e,kind:t},n.concepts=[...n.concepts,r]):r&&n.concepts.splice(s,1)}}setCardStageConceptValue(e,t,n){if(!this.currentCard||!this.currentCard.stages[e])return;const s=this.currentCard.stages[e],r=s.concepts.findIndex((e=>e.name==t));let i=-1!==r?s.concepts[r]:null;"PERMITTED"!==n?i?s.concepts[r]=Object.assign(Object.assign({},i),{kind:n}):(i={name:t,kind:n},s.concepts=[...s.concepts,i]):i&&s.concepts.splice(r,1)}setCardCommonLawValue(e,t){if(this.currentCard)for(const n of this.currentCard.stages){const s=n.laws.findIndex((t=>t.name==e));let r=-1!==s?n.laws[s]:null;"PERMITTED"!==t?r?n.laws[s]=Object.assign(Object.assign({},r),{kind:t}):(r={name:e,kind:t},n.laws=[...n.laws,r]):r&&n.laws.splice(s,1)}}setCardStageLawValue(e,t,n){if(!this.currentCard||!this.currentCard.stages[e])return;const s=this.currentCard.stages[e],r=s.laws.findIndex((e=>e.name==t));let i=-1!==r?s.laws[r]:null;"PERMITTED"!==n?i?s.laws[r]=Object.assign(Object.assign({},i),{kind:n}):(i={name:t,kind:n},s.laws=[...s.laws,i]):i&&s.laws.splice(r,1)}setCardStageSkillValue(e,t,n){if(!this.currentCard||!this.currentCard.stages[e])return;const s=this.currentCard.stages[e],r=s.skills.findIndex((e=>e.name==t));let i=-1!==r?s.skills[r]:null;"PERMITTED"!==n?i?s.skills[r]=Object.assign(Object.assign({},i),{kind:n}):(i={name:t,kind:n},s.skills=[...s.skills,i]):i&&s.skills.splice(r,1)}setCardCommonSkillValue(e,t){if(this.currentCard)for(const n of this.currentCard.stages){const s=n.skills.findIndex((t=>t.name==e));let r=-1!==s?n.laws[s]:null;"PERMITTED"!==t?r?n.skills[s]=Object.assign(Object.assign({},r),{kind:t}):(r={name:e,kind:t},n.skills=[...n.skills,r]):r&&n.laws.splice(s,1)}}setCardStageNumberOfQuestions(e,t){if(!this.currentCard||!this.currentCard.stages[e])return;const n=this.currentCard.stages[e];if(!t.match(/^\d*$/))return;const s=+t||1;n.numberOfQuestions=s}setCardSurveyEnabled(e){this.currentCard&&(this.currentCard.options.surveyOptions?this.currentCard.options.surveyOptions.enabled=e:this.currentCard.options.surveyOptions={enabled:e,surveyId:""})}setCardSurveyId(e){this.currentCard&&(this.currentCard.options.surveyOptions?this.currentCard.options.surveyOptions.surveyId=e:this.currentCard.options.surveyOptions={enabled:!0,surveyId:e})}setCardTags(e){this.currentCard&&(this.currentCard.tags=e)}setCardFlag(e,t){this.currentCard&&(this.currentCard.options[e]=t)}addStage(){if(!this.currentCard||!this.domains)return;const e=this.currentCard,t=new an(this.exerciseSettingsController,e,{numberOfQuestions:10,complexity:.5,laws:[],concepts:[],skills:[]});this.currentCard.stages.push(t)}removeStage(e){if(!this.currentCard)return;const t=this.currentCard.stages.length;e<0||e>=t||(this.currentCard.stages[e][Symbol.dispose](),this.currentCard.stages.splice(e,1))}};on=function(e,t,n,s){var r,i=arguments.length,a=i<3?t:null===s?s=Object.getOwnPropertyDescriptor(t,n):s;if("object"==typeof Reflect&&"function"==typeof Reflect.decorate)a=Reflect.decorate(e,t,n,s);else for(var o=e.length-1;o>=0;o--)(r=e[o])&&(a=(i<3?r(a):i>3?r(t,n,a):r(t,n))||a);return i>3&&a&&Object.defineProperty(t,n,a),a}([(0,i._G)(),sn(0,(0,i.WQ)(Ge)),sn(1,(0,i.WQ)(Se)),sn(2,(0,i.WQ)(S)),function(e,t){if("object"==typeof Reflect&&"function"==typeof Reflect.metadata)return Reflect.metadata(e,t)}("design:paramtypes",[Ge,Object,Object])],on);var ln=function(e,t,n,s){return new(n||(n=Promise))((function(r,i){function a(e){try{l(s.next(e))}catch(e){i(e)}}function o(e){try{l(s.throw(e))}catch(e){i(e)}}function l(e){var t;e.done?r(e.value):(t=e.value,t instanceof n?t:new n((function(e){e(t)}))).then(a,o)}l((s=s.apply(e,t||[])).next())}))};const cn=(0,We.PA)((()=>{var e,t,n,s,a,o,l;const[c]=(0,r.useState)((()=>i.kL.resolve(on))),{t:u}=(0,qe.Bd)();(0,r.useEffect)((()=>{ln(void 0,void 0,void 0,(function*(){yield c.loadExercises();const e=new URL(window.location.href).searchParams.get("exerciseId");e&&(yield c.loadExercise(Number.parseInt(e)))}))}),[]);const d=(0,r.useCallback)((()=>{ln(void 0,void 0,void 0,(function*(){yield c.createNewExecise()}))}),[null===(e=c.exercises)||void 0===e?void 0:e.length]);if("LOADING"===c.exercisesLoadStatus)return r.createElement(lt,null);const{user:m}=c;return m?r.createElement("div",{className:"container-fluid"},r.createElement("div",{className:"pt-1 pb-3"},r.createElement(it,{text:u("exercisesettings_title"),languageHint:u("language_header"),language:null!==(n=null===(t=c.user)||void 0===t?void 0:t.language)&&void 0!==n?n:"EN",onLanguageClicked:null,userHint:u("signedin_as_header"),user:m.displayName,userHref:null})),r.createElement("div",{className:"flex-xl-nowrap row"},r.createElement("div",{className:"col-xl-3 col-md-3 col-12 d-flex flex-column"},r.createElement("button",{type:"button",className:"btn btn-primary mb-3",onClick:d},"Create new"),r.createElement("ul",{className:"list-group"},null===(s=c.exercises)||void 0===s?void 0:s.map((e=>{var t;return r.createElement(Ye.N_,{key:e.id,className:"list-group-item "+(e.id===(null===(t=c.currentCard)||void 0===t?void 0:t.id)?"active":""),to:`?exerciseId=${e.id}`,onClick:()=>c.loadExercise(e.id),title:e.name},e.name.length>22?`${e.name.substring(0,22)}...`:e.name)})))),r.createElement("div",{className:"col-xl-9 col-md-9 col-12"},r.createElement(un,{store:c,card:c.currentCard,domains:null!==(a=c.domains)&&void 0!==a?a:[],backends:null!==(o=c.backends)&&void 0!==o?o:[],strategies:null!==(l=c.strategies)&&void 0!==l?l:[]})))):r.createElement(lt,null)})),un=(0,We.PA)((e=>{var t,n,s,i,a,o,l;const{card:c,domains:u,backends:d,strategies:m,store:p}=e,{t:g}=(0,qe.Bd)();if((0,r.useMemo)((()=>[g("exercisesettings_optDenied"),g("exercisesettings_optAllowed"),g("exercisesettings_optTarget")]),[null===(t=p.user)||void 0===t?void 0:t.language]),"EXERCISELOADING"===p.exercisesLoadStatus)return r.createElement(lt,{delay:200});if(null==c)return r.createElement("div",null,"No exercise selected");const h=u.find((e=>e.id===c.domainId)),v=null==h?void 0:h.laws.filter((e=>(e.bitflags&Me.TargetEnabled)>0)),f=null==h?void 0:h.concepts.filter((e=>(e.bitflags&Me.TargetEnabled)>0)),b=null==h?void 0:h.skills,E=c.stages[0].laws.reduce(((e,t)=>(e[t.name]=t,e)),{}),w=c.stages[0].concepts.reduce(((e,t)=>(e[t.name]=t,e)),{}),y=null==h?void 0:h.laws.filter((e=>!(e.bitflags&Me.TargetEnabled))),x=null==h?void 0:h.concepts.filter((e=>!(e.bitflags&Me.TargetEnabled))),C=[],S=m.find((e=>e.id===c.strategyId));return r.createElement("div",null,r.createElement("form",{className:"exercise-settings-form"},r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold",htmlFor:"exampleInputEmail1"},g("exercisesettings_name")),r.createElement("input",{value:c.name,type:"email",className:"form-control",id:"exampleInputEmail1","aria-describedby":"emailHelp",placeholder:"Enter email",onChange:e=>p.setCardName(e.target.value)})),r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold"},g("exercisesettings_domain")),r.createElement("select",{id:"domainId",className:"form-control","aria-describedby":"domainDescription",value:c.domainId,onChange:e=>p.setCardDomain(e.target.value),title:null==h?void 0:h.displayName},null==u?void 0:u.map((e=>{var t;return r.createElement("option",{value:e.id,title:null!==(t=e.description)&&void 0!==t?t:e.displayName},e.displayName)}))),r.createElement("small",{id:"domainDescription",className:"form-text text-muted"},null!==(n=null==h?void 0:h.description)&&void 0!==n?n:"")),r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold"},g("exercisesettings_strategy")),r.createElement("select",{id:"strategyId",className:"form-control","aria-describedby":"strategyDescription",value:c.strategyId,onChange:e=>p.setCardStrategy(e.target.value),title:null==S?void 0:S.displayName},null==m?void 0:m.map((e=>{var t;return r.createElement("option",{value:e.id,title:null!==(t=e.description)&&void 0!==t?t:e.displayName},e.displayName)}))),r.createElement("small",{id:"strategyDescription",className:"form-text text-muted"},null!==(s=null==S?void 0:S.description)&&void 0!==s?s:"")),r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold"},g("exercisesettings_qopt")),r.createElement("div",{className:"form-check"},r.createElement("input",{checked:c.options.forceNewAttemptCreationEnabled,onChange:e=>p.setCardFlag("forceNewAttemptCreationEnabled",e.target.checked),type:"checkbox",className:"form-check-input",id:"forceNewAttemptCreationEnabled"}),r.createElement("label",{className:"form-check-label",htmlFor:"forceNewAttemptCreationEnabled"},g("exercisesettings_qopt_forceAttCreation"))),r.createElement("div",{className:"form-check"},r.createElement("input",{checked:c.options.correctAnswerGenerationEnabled,onChange:e=>p.setCardFlag("correctAnswerGenerationEnabled",e.target.checked),type:"checkbox",className:"form-check-input",id:"correctAnswerGenerationEnabled"}),r.createElement("label",{className:"form-check-label",htmlFor:"correctAnswerGenerationEnabled"},g("exercisesettings_qopt_genCorAnsw"))),r.createElement("div",{className:"form-check"},r.createElement("input",{checked:c.options.newQuestionGenerationEnabled,onChange:e=>p.setCardFlag("newQuestionGenerationEnabled",e.target.checked),type:"checkbox",className:"form-check-input",id:"newQuestionGenerationEnabled"}),r.createElement("label",{className:"form-check-label",htmlFor:"newQuestionGenerationEnabled"},g("exercisesettings_qopt_forceShowGenNextQ"))),r.createElement("div",{className:"form-check"},r.createElement("input",{checked:c.options.supplementaryQuestionsEnabled,onChange:e=>p.setCardFlag("supplementaryQuestionsEnabled",e.target.checked),type:"checkbox",className:"form-check-input",id:"supplementaryQuestionsEnabled"}),r.createElement("label",{className:"form-check-label",htmlFor:"supplementaryQuestionsEnabled"},g("exercisesettings_qopt_supQ"))),r.createElement("div",{className:"form-check"},r.createElement("input",{checked:c.options.preferDecisionTreeBasedSupplementaryEnabled,onChange:e=>p.setCardFlag("preferDecisionTreeBasedSupplementaryEnabled",e.target.checked),type:"checkbox",className:"form-check-input",id:"preferDecisionTreeBasedSupplementaryEnabled"}),r.createElement("label",{className:"form-check-label",htmlFor:"preferDecisionTreeBasedSupplementaryEnabled"},g("exercisesettings_qopt_preferDTsup")))),r.createElement("div",{className:"form-group"},r.createElement("label",{htmlFor:"survOptions",className:"font-weight-bold"},g("exercisesettings_survey")),r.createElement("div",{className:"input-group mb-3"},r.createElement("div",{className:"input-group-prepend"},r.createElement("div",{className:"input-group-text"},r.createElement("input",{checked:null===(i=c.options.surveyOptions)||void 0===i?void 0:i.enabled,type:"checkbox","aria-label":"Checkbox for following text input",onChange:e=>p.setCardSurveyEnabled(e.target.checked)}))),r.createElement("input",{id:"survOptions",type:"text",value:null===(a=c.options.surveyOptions)||void 0===a?void 0:a.surveyId,className:"form-control","aria-label":"Text input with checkbox",disabled:!(null===(o=c.options.surveyOptions)||void 0===o?void 0:o.enabled),onChange:e=>p.setCardSurveyId(e.target.value)}))),r.createElement("div",{className:"form-group"},r.createElement("label",{htmlFor:"exTagsValues",className:"font-weight-bold"},g("exercisesettings_tags")),null==h?void 0:h.tags.map((e=>r.createElement("div",{className:"form-check"},r.createElement("input",{checked:c.tags.includes(e),onChange:e=>e.target.checked?p.setCardTags([...new Set([...c.tags,e.target.value])]):p.setCardTags([...c.tags.filter((t=>t!==e.target.value))]),type:"checkbox",value:e,className:"form-check-input",id:`tag_checkbox_${e}`}),r.createElement("label",{className:"form-check-label",htmlFor:`tag_checkbox_${e}`},e))))),(null==x?void 0:x.length)&&r.createElement("div",{className:"form-group"},r.createElement("div",null,r.createElement("label",{className:"font-weight-bold"},g("exercisesettings_commonConcepts"))),r.createElement("div",{className:"row"},r.createElement("div",{className:"col-md-12"},r.createElement("div",{className:"list-group list-group-flush"},r.createElement(mn,{id:"common_concepts",store:p,concepts:x,cardConcepts:w,onChange:(e,t,n)=>{p.setCardCommonConceptValue(e.name,t),n?p.setCardCommonConceptValue(n.name,"PERMITTED"):e.childs.forEach((e=>p.setCardCommonConceptValue(e.name,t)))}})))))||null,(null==y?void 0:y.length)&&r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold"},g("exercisesettings_commonLaws")),r.createElement("div",{className:"list-group list-group-flush"},r.createElement(pn,{id:"common_laws",store:p,laws:y,cardLaws:E,onChange:(e,t,n)=>{p.setCardCommonLawValue(e.name,t),n?p.setCardCommonLawValue(n.name,"PERMITTED"):e.childs.forEach((e=>p.setCardCommonLawValue(e.name,t)))}})))||null,(null==C?void 0:C.length)&&r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold"},g("exercisesettings_commonSkills")),r.createElement("div",{className:"list-group list-group-flush"},r.createElement(gn,{id:"common_skills",store:p,skills:C,cardSkills:E,onChange:(e,t,n)=>{p.setCardCommonSkillValue(e.name,t),n?p.setCardCommonSkillValue(n.name,"PERMITTED"):e.childs.forEach((e=>p.setCardCommonSkillValue(e.name,t)))}})))||null,r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold"},g("exercisesettings_stages")),r.createElement("div",{className:"list-group list-group-flush"},c.stages.map(((e,t,n)=>r.createElement(dn,{key:t,store:p,stage:e,stageIdx:t,showDeleteBtn:n.length>1,strategy:S,stageDomainConcepts:f,stageDomainSkills:b,stageDomainLaws:v}))))),c.stages.length<5&&(null==S?void 0:S.options.multiStagesEnabled)?r.createElement("div",{style:{marginTop:"-1rem"}},r.createElement("button",{type:"button",className:"btn btn-success",onClick:()=>p.addStage()},g("exercisesettings_addStage"))):null),(null===(l=p.user)||void 0===l?void 0:l.roles.includes("ADMIN"))&&r.createElement("div",{className:"mt-5"},r.createElement("button",{type:"button",className:"btn btn-primary",onClick:()=>p.saveCard()},g("exercisesettings_save")),r.createElement("button",{type:"button",className:"btn btn-primary ml-2",onClick:()=>p.saveCard().then((()=>{var e;return null===(e=window.open(`${window.location.origin}/pages/exercise?exerciseId=${c.id}`,"_blank"))||void 0===e?void 0:e.focus()}))},g("exercisesettings_saveNopen")),r.createElement("button",{type:"button",className:"btn btn-primary ml-2",onClick:()=>{var e;return null===(e=window.open(`${window.location.origin}/pages/exercise?exerciseId=${c.id}`,"_blank"))||void 0===e?void 0:e.focus()}},g("exercisesettings_open")),(null==S?void 0:S.options.multiStagesEnabled)&&r.createElement("button",{type:"button",className:"btn btn-primary ml-2",onClick:()=>{var e;return null===(e=window.open(`${window.location.origin}/pages/exercise?exerciseId=${c.id}&debug`,"_blank"))||void 0===e?void 0:e.focus()}},g("exercisesettings_genDebugAtt")))||null)})),dn=(0,We.PA)((e=>{var t,n;const{t:s}=(0,qe.Bd)(),{store:i,stage:a,strategy:o,stageIdx:l,showDeleteBtn:c,stageDomainConcepts:u,stageDomainLaws:d,stageDomainSkills:m}=e;if(!i.currentCard)throw new Error("card not set");const p=a.concepts.reduce(((e,t)=>(e[t.name]=t,e)),{}),g=a.laws.reduce(((e,t)=>(e[t.name]=t,e)),{}),h=a.skills.reduce(((e,t)=>(e[t.name]=t,e)),{});return r.createElement("div",{className:"card mb-3"},r.createElement("div",{className:"card-body"},r.createElement("div",{className:"form-group",style:{display:"flex",justifyContent:"space-between"}},r.createElement("div",null,s("exercisesettings_stageN",{stageNumber:l+1})),r.createElement("div",null,r.createElement("span",null,s("exercisesettings_questionsInBank"),": "),"IN_PROGRESS"===a.bankLoadingState||null===a.bankSearchResult?r.createElement(lt,{styleOverride:{width:"1rem",height:"1rem"},delay:0}):r.createElement("span",null,`${a.bankSearchResult.count} (${null===(t=a.bankSearchResult)||void 0===t?void 0:t.topRatedCount})`))),(null==o?void 0:o.options.multiStagesEnabled)&&r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold",htmlFor:`numberOfQuestions_stage${l}`},s("exercisesettings_stageN_qnumber")),r.createElement("input",{type:"text",className:"form-control",id:`numberOfQuestions_stage${l}`,value:a.numberOfQuestions,onChange:e=>i.setCardStageNumberOfQuestions(l,e.target.value)}))||null,r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold"},s("exercisesettings_qcomplexity")),r.createElement("div",{className:"d-flex"},r.createElement("input",{type:"range",className:"form-control-range",id:`complexity_stage${l}`,value:100*(null!==(n=a.complexity)&&void 0!==n?n:.5),onChange:e=>i.setCardStageComplexity(l,e.target.value)}),r.createElement("div",{className:"ml-2"},a.complexity.toFixed(2)))),u&&u.length>0&&r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold"},s("exercisesettings_stageN_concepts")),r.createElement("div",{className:"list-group list-group-flush"},r.createElement(mn,{id:`stage${l}_concepts`,store:i,concepts:u,cardConcepts:p,onChange:(e,t,n)=>{i.setCardStageConceptValue(l,e.name,t),n?i.setCardStageConceptValue(l,n.name,"PERMITTED"):e.childs.forEach((e=>i.setCardStageConceptValue(l,e.name,t)))}})))||null,d&&d.length>0&&r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold"},s("exercisesettings_stageN_laws")),r.createElement("div",{className:"list-group list-group-flush"},r.createElement(pn,{id:`stage${l}_laws`,store:i,laws:d,cardLaws:g,onChange:(e,t,n)=>{i.setCardStageLawValue(l,e.name,t),n?i.setCardStageLawValue(l,n.name,"PERMITTED"):e.childs.forEach((e=>i.setCardStageLawValue(l,e.name,t)))}})))||null,m&&m.length>0&&r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold"},s("exercisesettings_stageN_skills")),r.createElement("div",{className:"list-group list-group-flush"},r.createElement(gn,{id:`stage${l}_skills`,store:i,skills:m,cardSkills:h,onChange:(e,t,n)=>{i.setCardStageSkillValue(l,e.name,t),n?i.setCardStageSkillValue(l,n.name,"PERMITTED"):e.childs.forEach((e=>i.setCardStageSkillValue(l,e.name,t)))}})))||null,r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold"},s("exercisesettings_stageN_matchedQuestionExamples")),"IN_PROGRESS"===a.bankLoadingState&&r.createElement(lt,{styleOverride:{width:"1rem",height:"1rem"},delay:0})||r.createElement("div",{className:"list-group"},0===a.bankSearchResult.questions.length?r.createElement("div",{className:"list-group-item"},s("exercisesettings_noQuestionsFound")):a.bankSearchResult.questions.map(((e,t)=>r.createElement("div",{key:t,className:"list-group-item"},r.createElement("a",{target:"_blank",href:`${y}/pages/question?metadataId=${e.metadataId}`},e.name)))))),c&&r.createElement("div",{className:"d-flex justify-content-end"},r.createElement("button",{type:"button",className:"btn btn-danger",onClick:()=>i.removeStage(l)},s("exercisesettings_removeStage")))||null))})),mn=(0,We.PA)((e=>{var t;const{id:n,store:s,concepts:i,cardConcepts:a,onChange:o}=e,{t:l}=(0,qe.Bd)(),c=s.currentCard;if(!c)throw new Error("card not set");const u=(0,r.useMemo)((()=>[l("exercisesettings_optDenied"),l("exercisesettings_optAllowed"),l("exercisesettings_optTarget")]),[null===(t=s.user)||void 0===t?void 0:t.language]);return r.createElement(r.Fragment,null,i.map(((e,t)=>{var s;return r.createElement("div",{key:t,className:"list-group-item p-0 bg-transparent pt-2 pb-2"},r.createElement("div",null,r.createElement("div",{className:"d-flex flex-row align-items-center"},r.createElement(_t,{id:`concept_${n}_toggle_${c.id}_${e.name}_${t}`,selected:hn(null===(s=a[e.name])||void 0===s?void 0:s.kind),values:fn(e),valueStyles:[{backgroundColor:"#eb2828"},null,{backgroundColor:"#009700"}],displayNames:u,onChange:t=>o(e,vn(t))}),r.createElement("div",{style:{marginLeft:"15px"}},e.displayName))),e.childs.length>0&&r.createElement("ul",{className:""},e.childs.map(((s,i)=>{var l;return r.createElement(r.Fragment,null,r.createElement("li",{key:i,className:"d-flex flex-row align-items-centers mt-3"},r.createElement(_t,{id:`concept_${n}_toggle_${c.id}_${s.name}_${t}_${i}`,selected:hn(null===(l=a[s.name])||void 0===l?void 0:l.kind),values:fn(s),valueStyles:[{backgroundColor:"#eb2828"},null,{backgroundColor:"#009700"}],displayNames:u,onChange:t=>o(s,vn(t),e)}),r.createElement("div",{style:{marginLeft:"15px"}},s.displayName)))}))))})))})),pn=(0,We.PA)((e=>{var t;const{id:n,store:s,laws:i,cardLaws:a,onChange:o}=e,{t:l}=(0,qe.Bd)(),c=s.currentCard;if(!c)throw new Error("card not set");const u=(0,r.useMemo)((()=>[l("exercisesettings_optDenied"),l("exercisesettings_optAllowed"),l("exercisesettings_optTarget")]),[null===(t=s.user)||void 0===t?void 0:t.language]);return r.createElement(r.Fragment,null,i.map(((e,t)=>{var s;return r.createElement("div",{className:"list-group-item p-0 bg-transparent pt-2 pb-2",key:t},r.createElement("div",null,r.createElement("div",{className:"d-flex flex-row align-items-center"},r.createElement(_t,{id:`law_${n}_toggle_${c.id}_${t}`,selected:hn(null===(s=a[e.name])||void 0===s?void 0:s.kind),values:["Denied","Allowed","Target"],valueStyles:[{backgroundColor:"#eb2828"},null,{backgroundColor:"#009700"}],displayNames:u,onChange:t=>o(e,vn(t))}),r.createElement("div",{style:{marginLeft:"15px"}},e.displayName))),e.childs.length>0&&r.createElement("ul",{className:""},e.childs.map(((s,i)=>{var l;return r.createElement(r.Fragment,null,r.createElement("li",{key:i,className:"d-flex flex-row align-items-centers mt-3"},r.createElement(_t,{id:`law_${n}_toggle_${c.id}_${t}_${i}`,selected:hn(null===(l=a[s.name])||void 0===l?void 0:l.kind),values:["Denied","Allowed","Target"],valueStyles:[{backgroundColor:"#eb2828"},null,{backgroundColor:"#009700"}],displayNames:u,onChange:t=>o(s,vn(t),e)}),r.createElement("div",{style:{marginLeft:"15px"}},s.displayName)))}))))})))})),gn=(0,We.PA)((e=>{var t;const{id:n,store:s,skills:i,cardSkills:a,onChange:o}=e,{t:l}=(0,qe.Bd)(),c=s.currentCard;if(!c)throw new Error("card not set");const u=(0,r.useMemo)((()=>[l("exercisesettings_optDenied"),l("exercisesettings_optAllowed"),l("exercisesettings_optTarget")]),[null===(t=s.user)||void 0===t?void 0:t.language]);return r.createElement(r.Fragment,null,i.map(((e,t)=>{var s;return r.createElement("div",{className:"list-group-item p-0 bg-transparent pt-2 pb-2",key:t},r.createElement("div",null,r.createElement("div",{className:"d-flex flex-row align-items-center"},r.createElement(_t,{id:`skill_${n}_toggle_${c.id}_${t}`,selected:hn(null===(s=a[e.name])||void 0===s?void 0:s.kind),values:["Denied","Allowed","Target"],valueStyles:[{backgroundColor:"#eb2828"},null,{backgroundColor:"#009700"}],displayNames:u,onChange:t=>o(e,vn(t))}),r.createElement("div",{style:{marginLeft:"15px"}},e.displayName))),e.childs.length>0&&r.createElement("ul",{className:""},e.childs.map(((s,i)=>{var l;return r.createElement(r.Fragment,null,r.createElement("li",{key:i,className:"d-flex flex-row align-items-centers mt-3"},r.createElement(_t,{id:`skill_${n}_toggle_${c.id}_${t}_${i}`,selected:hn(null===(l=a[s.name])||void 0===l?void 0:l.kind),values:["Denied","Allowed","Target"],valueStyles:[{backgroundColor:"#eb2828"},null,{backgroundColor:"#009700"}],displayNames:u,onChange:t=>o(s,vn(t),e)}),r.createElement("div",{style:{marginLeft:"15px"}},s.displayName)))}))))})))}));function hn(e){return"FORBIDDEN"===e?"Denied":"TARGETED"===e?"Target":"Allowed"}function vn(e){return"Denied"===e?"FORBIDDEN":"Target"===e?"TARGETED":"PERMITTED"}function fn(e){return(e.bitflags&Me.TargetEnabled)>0?["Denied","Allowed","Target"]:["Denied","Allowed"]}const bn={concepts:["literal","variable","arithmetic operator","assignment operator","comparison operator","logical operator","bitwise operator","array access operator","pointer operator","function call"],laws:["Expression contains two operators with different precedence in a row","Expression contains two operators with the same precedence and left associativity in a row","Expression contains two operators with the same precedence and right associativity in a row","Expression contains operators inside parentheses or another operator","Expression contains operators evaluating their operands in a strict order"]},En=(0,We.PA)((()=>r.createElement("div",{className:"container-fluid",style:{color:"black",fontSize:"18px"}},r.createElement("div",{className:"row"},r.createElement("div",{className:"col-md-8"},r.createElement("form",null,r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold"},"Domain"),r.createElement("select",{className:"form-control"},r.createElement("option",null,"Order of Expression Evaluation"))),r.createElement("div",{className:"row"},r.createElement("div",{className:"col-md-6"},r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold"},"Question complexity"),r.createElement("div",null,r.createElement("input",{type:"range",className:"form-control-range",id:"formControlRange1"})))),r.createElement("div",{className:"col-md-6"},r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold"},"Answer length"),r.createElement("div",null,r.createElement("input",{type:"range",className:"form-control-range",id:"formControlRange1"}))))),r.createElement("div",{className:"form-group"},r.createElement("div",null,r.createElement("label",{className:"font-weight-bold"},"Concepts")),r.createElement("div",{className:"row"},r.createElement("div",{className:"col-md-6"},bn.concepts.filter(((e,t)=>t%2==0)).map(((e,t)=>r.createElement("div",{className:"d-flex flex-row align-items-center",style:{marginBottom:"10px"}},r.createElement(_t,{id:`concept_toggle_${e}_${t}`,selected:"Allowed",values:["Denied","Allowed","Target"],valueStyles:[{backgroundColor:"#eb2828"},null,{backgroundColor:"#009700"}],onChange:e=>0}),r.createElement("div",{style:{marginLeft:"15px"}},e))))),r.createElement("div",{className:"col-md-6"},bn.concepts.filter(((e,t)=>t%2!=0)).map(((e,t)=>r.createElement("div",{className:"d-flex flex-row align-items-center",style:{marginBottom:"10px"}},r.createElement(_t,{id:`concept_toggle_${e}_${t}`,selected:"Allowed",values:["Denied","Allowed","Target"],valueStyles:[{backgroundColor:"#eb2828"},null,{backgroundColor:"#009700"}],onChange:e=>0}),r.createElement("div",{style:{marginLeft:"15px"}},e))))))),r.createElement("div",{className:"form-group"},r.createElement("label",{className:"font-weight-bold"},"Laws"),bn.laws.map(((e,t)=>r.createElement("div",{className:"d-flex flex-row align-items-start justify-content-start",style:{marginBottom:"10px"}},r.createElement("div",null,r.createElement(_t,{id:`law_toggle_${t}`,selected:"Allowed",values:["Denied","Allowed","Target"],valueStyles:[{backgroundColor:"#eb2828"},null,{backgroundColor:"#009700"}],onChange:e=>0})),r.createElement("div",{style:{marginLeft:"15px"}},e))))))))))),wn=(0,We.PA)((()=>{const[e]=(0,r.useState)((()=>i.kL.resolve(ie)));return(0,r.useEffect)((()=>{const t=new URLSearchParams(window.location.search).get("metadataId");if(!t||Number.isNaN(Number(t)))throw new Error("no metadataId provided in the URL");e.generateQuestionByMetadata(+t)}),[e]),r.createElement(r.Fragment,null,r.createElement(Vt,{store:e,showExtendedFeedback:!0}),r.createElement("div",{className:"mt-3"},r.createElement(ut,{explicitQuestionStore:e})))})),yn=document.getElementById("root");(0,s.H)(yn).render(r.createElement((()=>r.createElement("div",{className:"container comp-ph-container"},r.createElement(Ye.Kd,null,r.createElement(ze.BV,null,r.createElement(ze.qh,{path:"/pages/statistics",element:r.createElement(Zt,null)}),r.createElement(ze.qh,{path:"/pages/exercise",element:r.createElement(Kt,null)}),r.createElement(ze.qh,{path:"/pages/exercise-settings",element:r.createElement(cn,null)}),r.createElement(ze.qh,{path:"/pages/strategy-settings",element:r.createElement(En,null)}),r.createElement(ze.qh,{path:"/pages/survey",element:r.createElement(nn,null)}),r.createElement(ze.qh,{path:"/pages/question",element:r.createElement(wn,null)}),r.createElement(ze.qh,{path:"/pages/exercises-list",element:r.createElement(en,null)}),r.createElement(ze.qh,{path:"/",element:r.createElement(ze.C5,{to:"/pages/exercise-settings",replace:!0})}))))),null))}},r={};function i(e){var t=r[e];if(void 0!==t)return t.exports;var n=r[e]={exports:{}};return s[e].call(n.exports,n,n.exports,i),n.exports}i.m=s,e=[],i.O=(t,n,s,r)=>{if(!n){var a=1/0;for(u=0;u<e.length;u++){for(var[n,s,r]=e[u],o=!0,l=0;l<n.length;l++)(!1&r||a>=r)&&Object.keys(i.O).every((e=>i.O[e](n[l])))?n.splice(l--,1):(o=!1,r<a&&(a=r));if(o){e.splice(u--,1);var c=s();void 0!==c&&(t=c)}}return t}r=r||0;for(var u=e.length;u>0&&e[u-1][2]>r;u--)e[u]=e[u-1];e[u]=[n,s,r]},i.n=e=>{var t=e&&e.__esModule?()=>e.default:()=>e;return i.d(t,{a:t}),t},n=Object.getPrototypeOf?e=>Object.getPrototypeOf(e):e=>e.__proto__,i.t=function(e,s){if(1&s&&(e=this(e)),8&s)return e;if("object"==typeof e&&e){if(4&s&&e.__esModule)return e;if(16&s&&"function"==typeof e.then)return e}var r=Object.create(null);i.r(r);var a={};t=t||[null,n({}),n([]),n(n)];for(var o=2&s&&e;"object"==typeof o&&!~t.indexOf(o);o=n(o))Object.getOwnPropertyNames(o).forEach((t=>a[t]=()=>e[t]));return a.default=()=>e,i.d(r,a),r},i.d=(e,t)=>{for(var n in t)i.o(t,n)&&!i.o(e,n)&&Object.defineProperty(e,n,{enumerable:!0,get:t[n]})},i.g=function(){if("object"==typeof globalThis)return globalThis;try{return this||new Function("return this")()}catch(e){if("object"==typeof window)return window}}(),i.o=(e,t)=>Object.prototype.hasOwnProperty.call(e,t),i.r=e=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},(()=>{var e={818:0};i.O.j=t=>0===e[t];var t=(t,n)=>{var s,r,[a,o,l]=n,c=0;if(a.some((t=>0!==e[t]))){for(s in o)i.o(o,s)&&(i.m[s]=o[s]);if(l)var u=l(i)}for(t&&t(n);c<a.length;c++)r=a[c],i.o(e,r)&&e[r]&&e[r][0](),e[r]=0;return i.O(u)},n=self.webpackChunkcomp_prehension=self.webpackChunkcomp_prehension||[];n.forEach(t.bind(null,0)),n.push=t.bind(null,n.push.bind(n))})();var a=i.O(void 0,[121],(()=>i(558)));a=i.O(a)})();
